TsQtAWS refactor
