+++ sequence/CMakeLists.txt
@@ -62,8 +62,8 @@ if (TOPAS_USE_QT_DETECTED AND NOT TOPAS_USE_QT)
 endif()
 
 if (TOPAS_USE_QT6)
-	list(APPEND TOPAS_SEQUENCE_SRC TsQt6.cc TsQtAWS.cc)
-	list(APPEND TOPAS_SEQUENCE_HEADERS TsQt6.hh TsQtAWS.hh)
+	list(APPEND TOPAS_SEQUENCE_SRC TsQt6.cc TsQtAWS.cc TsQtAWSUtils.cc TsQtAWSPages.cc)
+	list(APPEND TOPAS_SEQUENCE_HEADERS TsQt6.hh TsQtAWS.hh TsQtAWSUtils.hh)
 elseif (TOPAS_USE_QT)
 	list(APPEND TOPAS_SEQUENCE_SRC TsQt5.cc)
 	list(APPEND TOPAS_SEQUENCE_HEADERS TsQt5.hh)
