+++ sequence/TsQtAWS.cc
@@ -30,11 +30,11 @@
 #ifdef G4UI_USE_QT
 
 #include "TsQtAWS.hh"
+#include "TsQtAWSUtils.hh"
 
 #include "TsParameterManager.hh"
 
 #include <QCoreApplication>
-#include <QCheckBox>
 #include <QComboBox>
 #include <QDate>
 #include <QDateTime>
@@ -78,880 +78,7 @@
 #include <QTextDocument>
 #include <functional>
 
-namespace {
-
-const char* kFileCompute = "batch-compute-env.json";
-const char* kFileJobDef = "batch-job-definition.json";
-const char* kFileJobDefPost = "batch-postP-job-definition.json";
-const char* kFileSubmit = "topas_submit.sh";
-const char* kFilePost = "postProcess_submit.sh";
-
-QString AwsFieldTooltipForField(const QString& fieldId)
-{
-    static const QMap<QString, QString> kFieldTooltips = {
-        { QStringLiteral("computeEnvironmentName"), QStringLiteral("The name for your compute environment. It can be up to 128 characters long. It can contain uppercase \nand lowercase letters, numbers, hyphens (-), and underscores (_).") },
-        { QStringLiteral("type"), QStringLiteral("The type of the compute environment: MANAGED or UNMANAGED.") },
-        { QStringLiteral("state"), QStringLiteral("The state of the compute environment. If the state is ENABLED, then the compute environment accepts jobs from a queue and \ncan scale out automatically based on queues.") },
-        { QStringLiteral("serviceRole"), QStringLiteral("The full Amazon Resource Name (ARN) of the IAM role that allows AWS Batch to make calls to other AWS services \non your behalf.") },
-        { QStringLiteral("computeResources"), QStringLiteral("Details about the compute resources managed by the compute environment. This parameter is required for \nmanaged compute environments.") },
-        { QStringLiteral("computeResources.type"), QStringLiteral("The type of compute environment: EC2, SPOT, FARGATE, or FARGATE_SPOT.") },
-        { QStringLiteral("computeResources.minvCpus"), QStringLiteral("The minimum number of vCPUs that an environment should maintain (even if the compute \nenvironment is DISABLED).") },
-        { QStringLiteral("computeResources.maxvCpus"), QStringLiteral("The maximum number of Amazon EC2 vCPUs that an environment can reach.") },
-        { QStringLiteral("computeResources.desiredvCpus"), QStringLiteral("The desired number of vCPUS in the compute environment. AWS Batch modifies this value between the \nminimum and maximum values based on job queue demand.") },
-        { QStringLiteral("computeResources.instanceTypes"), QStringLiteral("The instances types that can be launched. By choosing 'optimal', AWS Batch will automatically selects \nan x86_65 based instance that best matches the resource demands of the job queue.") },
-        { QStringLiteral("computeResources.subnets"), QStringLiteral("A VPC (Virtual Private Cloud) is a private network inside AWS attached to your AWS account. A subnet is a \nrange of IP addresses inside your VPC. Batch needs at least one subnet so it has a place to start the EC2 instances that run your jobs") },
-        { QStringLiteral("computeResources.securityGroupIds"), QStringLiteral("A security group controls what network traffic is allowed in and out of the instances.") },
-        { QStringLiteral("computeResources.instanceRole"), QStringLiteral("The Amazon ECS instance profile applied to Amazon EC2 instances in a compute environment.") },
-        { QStringLiteral("jobDefinitionName"), QStringLiteral("The name of the job definition.") },
-        { QStringLiteral("platformCapabilities"), QStringLiteral("The platform capabilities required by the job definition. If no value is specified, \nit defaults to using the EC2 compute environment.") },
-        { QStringLiteral("containerProperties"), QStringLiteral("These properties to describe the container that's launched as part of a job.") },
-        { QStringLiteral("containerProperties.image"), QStringLiteral("The image used to start a container. This should be your own/the official \nTOPAS/TOPAS-nBio images on AWS ECR or Docker Hub.") },
-        { QStringLiteral("containerProperties.environment"), QStringLiteral("The environment variables to pass to a container.") },
-        { QStringLiteral("containerProperties.resourceRequirements"), QStringLiteral("The type and amount of resources to assign to a container. \nThe supported resources include GPU, MEMORY, and VCPU.") },
-        { QStringLiteral("containerProperties.logConfiguration"), QStringLiteral("The logging system to use.") },
-        { QStringLiteral("containerProperties.logConfiguration.logDriver"), QStringLiteral("Since CloudWatch logs are used, the awslogs driver is needed \nto send stdout/stderr logs directly to CloudWatch") },
-        { QStringLiteral("containerProperties.logConfiguration.options"), QStringLiteral("The configuration options to send to the log driver.") },
-        { QStringLiteral("containerProperties.logConfiguration.options.awslogs-group"), QStringLiteral("The CloudWatch log group to send logs to. This was specified in the previous step.") },
-        { QStringLiteral("containerProperties.logConfiguration.options.awslogs-region"), QStringLiteral("The AWS region of the account.") },
-        { QStringLiteral("retryStrategy"), QStringLiteral("The retry strategy to use for failed jobs that are submitted with this job definition.") },
-        { QStringLiteral("retryStrategy.attempts"), QStringLiteral("The number of times to move a job to the RUNNABLE status. You can specify between 1 and 10 attempts.") },
-    };
-
-    if (kFieldTooltips.contains(fieldId))
-        return kFieldTooltips.value(fieldId);
-    return QStringLiteral("Please refer to the AWS documentation for more information.");
-}
-
-QString AwsConfigGroupForProfile(const QString& profileName)
-{
-    if (profileName == QLatin1String("default"))
-        return QStringLiteral("default");
-    return QStringLiteral("profile ") + profileName;
-}
-
-void ApplyAwsSharedConfig(const QString& configPath, QString* profile, QString* region)
-{
-    if (!profile || !region || !QFileInfo::exists(configPath))
-        return;
-
-    QSettings cfg(configPath, QSettings::IniFormat);
-    const QStringList groups = cfg.childGroups();
-
-    if (profile->isEmpty()) {
-        QStringList named;
-        const QString pref = QStringLiteral("profile ");
-        for (const QString& g : groups) {
-            if (g.startsWith(pref))
-                named.append(g.mid(pref.size()));
-        }
-        if (named.size() == 1)
-            *profile = named[0];
-        else if (named.isEmpty() && groups.contains(QStringLiteral("default")))
-            *profile = QStringLiteral("default");
-    }
-
-    if (region->isEmpty() && !profile->isEmpty()) {
-        cfg.beginGroup(AwsConfigGroupForProfile(*profile));
-        const QString reg = cfg.value(QStringLiteral("region")).toString().trimmed();
-        cfg.endGroup();
-        if (!reg.isEmpty())
-            *region = reg;
-    }
-}
-
-enum class JsonArrayShape
-{
-    Primitive,
-    FlatObject,
-    Complex
-};
-
-JsonArrayShape ClassifyArrayShape(const QJsonArray& arr)
-{
-    if (arr.isEmpty())
-        return JsonArrayShape::Primitive;
-
-    bool hasObject = false;
-    for (const QJsonValue& v : arr) {
-        if (v.isArray())
-            return JsonArrayShape::Complex;
-        if (v.isObject()) {
-            hasObject = true;
-            const QJsonObject obj = v.toObject();
-            for (auto it = obj.constBegin(); it != obj.constEnd(); ++it) {
-                if (it.value().isObject() || it.value().isArray())
-                    return JsonArrayShape::Complex;
-            }
-        }
-    }
-    return hasObject ? JsonArrayShape::FlatObject : JsonArrayShape::Primitive;
-}
-
-QStringList KeysFromObjectIteration(const QJsonObject& obj)
-{
-    QStringList keys;
-    for (auto it = obj.constBegin(); it != obj.constEnd(); ++it)
-        keys.append(it.key());
-    return keys;
-}
-
-void AddFlatObjectArrayRows(const QString& path, const QJsonArray& arr, QVBoxLayout* parentLayout, QWidget* wizard,
-                            QMap<QString, QLineEdit*>* scalarEdits, QMap<QString, QJsonValue::Type>* scalarTypes,
-                            QMap<QString, QVBoxLayout*>* flatRowsLayouts, QMap<QString, QStringList>* flatKeys,
-                            QMap<QString, QList<QWidget*>>* flatRowWidgets, const std::function<void()>& markNeedsApply)
-{
-    auto appendRow = [path, wizard, scalarEdits, scalarTypes, markNeedsApply](
-                         const QStringList& orderedKeys, const QJsonObject& itemObj,
-                         QVBoxLayout* targetRowsLayout, QList<QWidget*>* rowsStore) {
-        if (!targetRowsLayout || !rowsStore)
-            return;
-        const int rowIndex = rowsStore->size();
-        QWidget* entryWidget = new QWidget();
-        QGridLayout* grid = new QGridLayout(entryWidget);
-        grid->setContentsMargins(24, 0, 0, 0);
-        grid->setHorizontalSpacing(12);
-        grid->setVerticalSpacing(6);
-        for (int col = 0; col < orderedKeys.size(); ++col) {
-            const QString& key = orderedKeys[col];
-            QLabel* fieldLabel = new QLabel(QString("<b>%1</b>").arg(key));
-            fieldLabel->setTextFormat(Qt::RichText);
-            fieldLabel->setToolTip(AwsFieldTooltipForField(QStringLiteral("%1.<row>.%2").arg(path).arg(key)));
-            grid->addWidget(fieldLabel, 0, col);
-
-            const QJsonValue v = itemObj.value(key);
-            QLineEdit* edit = new QLineEdit(v.toVariant().toString());
-            edit->setFixedHeight(22);
-            edit->setCursorPosition(0);
-            QObject::connect(edit, &QLineEdit::textEdited, wizard, [markNeedsApply](const QString&) { markNeedsApply(); });
-            grid->addWidget(edit, 1, col);
-            const QString fieldPath = QString("%1.%2.%3").arg(path).arg(rowIndex).arg(key);
-            (*scalarEdits)[fieldPath] = edit;
-            QJsonValue::Type t = v.type();
-            if (t == QJsonValue::Undefined || t == QJsonValue::Null)
-                t = QJsonValue::String;
-            (*scalarTypes)[fieldPath] = t;
-        }
-        rowsStore->append(entryWidget);
-        const int insertAt = std::max(0, targetRowsLayout->count() - 1); // keep +/- at bottom
-        targetRowsLayout->insertWidget(insertAt, entryWidget);
-    };
-
-    auto remapScalarPathsForArray = [path, scalarEdits, scalarTypes](const QStringList& orderedKeys, QList<QWidget*>* rowsStore) {
-        if (!rowsStore)
-            return;
-        const QString prefix = path + ".";
-        QStringList toRemove;
-        for (auto it = scalarEdits->constBegin(); it != scalarEdits->constEnd(); ++it) {
-            if (it.key().startsWith(prefix))
-                toRemove.append(it.key());
-        }
-        QMap<QString, QJsonValue::Type> keyTypes;
-        for (const QString& key : orderedKeys) {
-            keyTypes[key] = QJsonValue::String;
-            for (auto it = scalarTypes->constBegin(); it != scalarTypes->constEnd(); ++it) {
-                if (it.key().startsWith(prefix) && it.key().endsWith("." + key)) {
-                    keyTypes[key] = it.value();
-                    break;
-                }
-            }
-        }
-        for (const QString& k : toRemove) {
-            scalarEdits->remove(k);
-            scalarTypes->remove(k);
-        }
-        for (int row = 0; row < rowsStore->size(); ++row) {
-            QWidget* rowWidget = (*rowsStore)[row];
-            QGridLayout* grid = qobject_cast<QGridLayout*>(rowWidget ? rowWidget->layout() : nullptr);
-            if (!grid)
-                continue;
-            for (int col = 0; col < orderedKeys.size(); ++col) {
-                const QString& key = orderedKeys[col];
-                QLayoutItem* item = grid->itemAtPosition(1, col);
-                QLineEdit* edit = item ? qobject_cast<QLineEdit*>(item->widget()) : nullptr;
-                if (!edit)
-                    continue;
-                const QString fieldPath = QString("%1.%2.%3").arg(path).arg(row).arg(key);
-                (*scalarEdits)[fieldPath] = edit;
-                (*scalarTypes)[fieldPath] = keyTypes.value(key, QJsonValue::String);
-            }
-        }
-    };
-
-    QWidget* rowsWidget = new QWidget();
-    QVBoxLayout* rowsLayout = new QVBoxLayout(rowsWidget);
-    rowsLayout->setContentsMargins(24, 0, 0, 0);
-    rowsLayout->setSpacing(9);
-    QStringList orderedKeys;
-    if (!arr.isEmpty() && arr[0].isObject())
-        orderedKeys = KeysFromObjectIteration(arr[0].toObject());
-    if (orderedKeys.isEmpty())
-        orderedKeys.append("value");
-    (*flatRowsLayouts)[path] = rowsLayout;
-    (*flatKeys)[path] = orderedKeys;
-    (*flatRowWidgets)[path].clear();
-
-    for (const QJsonValue& v : arr)
-        appendRow(orderedKeys, v.toObject(), rowsLayout, &((*flatRowWidgets)[path]));
-
-    QHBoxLayout* btnRow = new QHBoxLayout();
-    btnRow->setSpacing(8);
-    QPushButton* addBtn = new QPushButton("+");
-    addBtn->setFixedWidth(30);
-    QPushButton* removeBtn = new QPushButton("-");
-    removeBtn->setFixedWidth(30);
-    btnRow->addWidget(addBtn);
-    btnRow->addWidget(removeBtn);
-    btnRow->addStretch(1);
-    QObject::connect(addBtn, &QPushButton::clicked, wizard, [=]() {
-        const QStringList keys = flatKeys->value(path);
-        QVBoxLayout* targetRowsLayout = flatRowsLayouts->value(path, nullptr);
-        if (!targetRowsLayout)
-            return;
-        QDialog dialog(wizard);
-        dialog.setWindowTitle("Add value");
-        QVBoxLayout* v = new QVBoxLayout(&dialog);
-        QGridLayout* g = new QGridLayout();
-        QList<QLineEdit*> inputs;
-        for (int i = 0; i < keys.size(); ++i) {
-            QLabel* lbl = new QLabel(QString("%1:").arg(keys[i]));
-            QLineEdit* in = new QLineEdit();
-            in->setFixedHeight(22);
-            g->addWidget(lbl, i, 0);
-            g->addWidget(in, i, 1);
-            inputs.append(in);
-        }
-        v->addLayout(g);
-        QDialogButtonBox* box = new QDialogButtonBox();
-        QPushButton* add = box->addButton("Add", QDialogButtonBox::AcceptRole);
-        add->setStyleSheet(
-            "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-            "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-            "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
-        box->addButton("Cancel", QDialogButtonBox::RejectRole);
-        v->addWidget(box);
-        QObject::connect(box, &QDialogButtonBox::accepted, &dialog, &QDialog::accept);
-        QObject::connect(box, &QDialogButtonBox::rejected, &dialog, &QDialog::reject);
-        if (dialog.exec() != QDialog::Accepted)
-            return;
-        QJsonObject obj;
-        for (int i = 0; i < keys.size(); ++i)
-            obj.insert(keys[i], inputs[i]->text().trimmed());
-        appendRow(keys, obj, targetRowsLayout, &((*flatRowWidgets)[path]));
-        markNeedsApply();
-    });
-    QObject::connect(removeBtn, &QPushButton::clicked, wizard, [=]() {
-        QList<QWidget*>& rows = (*flatRowWidgets)[path];
-        if (rows.isEmpty())
-            return;
-        const QStringList keys = flatKeys->value(path);
-        QDialog dialog(wizard);
-        dialog.setWindowTitle("Remove value");
-        QVBoxLayout* v = new QVBoxLayout(&dialog);
-        v->addWidget(new QLabel("Select entry to remove:"));
-        QComboBox* combo = new QComboBox();
-        QVector<int> indices;
-        for (int row = 0; row < rows.size(); ++row) {
-            QWidget* rowWidget = rows[row];
-            QGridLayout* grid = qobject_cast<QGridLayout*>(rowWidget ? rowWidget->layout() : nullptr);
-            if (!grid)
-                continue;
-            QString summary;
-            for (int col = 0; col < keys.size(); ++col) {
-                QLayoutItem* item = grid->itemAtPosition(1, col);
-                QLineEdit* edit = item ? qobject_cast<QLineEdit*>(item->widget()) : nullptr;
-                if (!edit)
-                    continue;
-                if (!summary.isEmpty())
-                    summary += ", ";
-                summary += QString("%1=%2").arg(keys[col], edit->text());
-            }
-            combo->addItem(summary.isEmpty() ? QString("Entry %1").arg(row + 1) : summary);
-            indices.append(row);
-        }
-        v->addWidget(combo);
-        QDialogButtonBox* box = new QDialogButtonBox();
-        QPushButton* remove = box->addButton("Remove", QDialogButtonBox::AcceptRole);
-        remove->setEnabled(combo->count() > 0);
-        remove->setStyleSheet(
-            "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-            "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-            "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
-        box->addButton("Cancel", QDialogButtonBox::RejectRole);
-        v->addWidget(box);
-        QObject::connect(box, &QDialogButtonBox::accepted, &dialog, &QDialog::accept);
-        QObject::connect(box, &QDialogButtonBox::rejected, &dialog, &QDialog::reject);
-        if (dialog.exec() != QDialog::Accepted || combo->currentIndex() < 0)
-            return;
-        const int idx = indices[combo->currentIndex()];
-        if (idx < 0 || idx >= rows.size())
-            return;
-        QWidget* victim = rows.takeAt(idx);
-        for (const QString& key : keys) {
-            const QString fieldPath = QString("%1.%2.%3").arg(path).arg(idx).arg(key);
-            (*scalarEdits).remove(fieldPath);
-            (*scalarTypes).remove(fieldPath);
-        }
-        delete victim;
-        remapScalarPathsForArray(keys, &rows);
-        markNeedsApply();
-    });
-    rowsLayout->addLayout(btnRow);
-    parentLayout->addWidget(rowsWidget);
-}
-
-int SkipWs(const QString& s, int pos)
-{
-    while (pos < s.size() && s[pos].isSpace())
-        ++pos;
-    return pos;
-}
-
-bool ParseJsonStringAt(const QString& s, int quotePos, QString* out, int* endPos)
-{
-    if (!out || !endPos || quotePos < 0 || quotePos >= s.size() || s[quotePos] != QChar('"'))
-        return false;
-    QString result;
-    bool esc = false;
-    int i = quotePos + 1;
-    for (; i < s.size(); ++i) {
-        const QChar c = s[i];
-        if (esc) {
-            result.append(c);
-            esc = false;
-            continue;
-        }
-        if (c == QChar('\\')) {
-            esc = true;
-            continue;
-        }
-        if (c == QChar('"')) {
-            *out = result;
-            *endPos = i + 1;
-            return true;
-        }
-        result.append(c);
-    }
-    return false;
-}
-
-int FindRootObjectStartByIndex(const QString& jsonText, int targetIndex)
-{
-    int pos = SkipWs(jsonText, 0);
-    if (pos >= jsonText.size() || jsonText[pos] != QChar('['))
-        return -1;
-    bool inString = false;
-    bool esc = false;
-    int depthObj = 0;
-    int depthArr = 0;
-    int currentObjectIdx = -1;
-    for (int i = pos; i < jsonText.size(); ++i) {
-        const QChar c = jsonText[i];
-        if (inString) {
-            if (esc)
-                esc = false;
-            else if (c == QChar('\\'))
-                esc = true;
-            else if (c == QChar('"'))
-                inString = false;
-            continue;
-        }
-        if (c == QChar('"')) {
-            inString = true;
-            continue;
-        }
-        if (c == QChar('[')) {
-            ++depthArr;
-            continue;
-        }
-        if (c == QChar(']')) {
-            --depthArr;
-            if (depthArr <= 0)
-                break;
-            continue;
-        }
-        if (c == QChar('{')) {
-            if (depthArr == 1 && depthObj == 0) {
-                ++currentObjectIdx;
-                if (currentObjectIdx == targetIndex)
-                    return i;
-            }
-            ++depthObj;
-            continue;
-        }
-        if (c == QChar('}')) {
-            --depthObj;
-            continue;
-        }
-    }
-    return -1;
-}
-
-
-QStringList ExtractFirstLevelObjectKeys(const QString& jsonText, int objectStartPos)
-{
-    QStringList keys;
-    if (objectStartPos < 0 || objectStartPos >= jsonText.size() || jsonText[objectStartPos] != QChar('{'))
-        return keys;
-    bool inString = false;
-    bool esc = false;
-    int depthObj = 0;
-    int depthArr = 0;
-    for (int i = objectStartPos; i < jsonText.size(); ++i) {
-        const QChar c = jsonText[i];
-        if (inString) {
-            if (esc)
-                esc = false;
-            else if (c == QChar('\\'))
-                esc = true;
-            else if (c == QChar('"'))
-                inString = false;
-            continue;
-        }
-        if (c == QChar('"')) {
-            if (depthObj == 1 && depthArr == 0) {
-                QString key;
-                int after = i;
-                if (ParseJsonStringAt(jsonText, i, &key, &after)) {
-                    const int colonPos = SkipWs(jsonText, after);
-                    if (colonPos < jsonText.size() && jsonText[colonPos] == QChar(':')) {
-                        keys.append(key);
-                        i = after - 1;
-                        continue;
-                    }
-                }
-            }
-            inString = true;
-            continue;
-        }
-        if (c == QChar('{')) {
-            ++depthObj;
-            continue;
-        }
-        if (c == QChar('}')) {
-            --depthObj;
-            if (depthObj == 0)
-                break;
-            continue;
-        }
-        if (c == QChar('[')) {
-            ++depthArr;
-            continue;
-        }
-        if (c == QChar(']')) {
-            --depthArr;
-            continue;
-        }
-    }
-    return keys;
-}
-
-QStringList ExtractTopLevelJobDefKeys(const QString& rawJson, const QString& path)
-{
-    const bool isRootObject = path.isEmpty();
-    bool isRootArrayIndex = false;
-    const int rootIndex = path.toInt(&isRootArrayIndex);
-    if (!isRootObject && !isRootArrayIndex)
-        return QStringList();
-    if (isRootObject) {
-        const int start = rawJson.indexOf('{');
-        return ExtractFirstLevelObjectKeys(rawJson, start);
-    }
-    const int start = FindRootObjectStartByIndex(rawJson, rootIndex);
-    return ExtractFirstLevelObjectKeys(rawJson, start);
-}
-
-QString JsonEscaped(const QString& s)
-{
-    QString out;
-    out.reserve(s.size() + 8);
-    for (QChar c : s) {
-        switch (c.unicode()) {
-        case '\"': out += "\\\""; break;
-        case '\\': out += "\\\\"; break;
-        case '\b': out += "\\b"; break;
-        case '\f': out += "\\f"; break;
-        case '\n': out += "\\n"; break;
-        case '\r': out += "\\r"; break;
-        case '\t': out += "\\t"; break;
-        default: out += c; break;
-        }
-    }
-    return out;
-}
-
-QString Indent(int n)
-{
-    return QString(n, QChar(' '));
-}
-
-QString SerializeJsonValueOrdered(const QJsonValue& value, int indent, const QStringList* preferredKeys = nullptr);
-
-QString SerializeJsonObjectOrdered(const QJsonObject& obj, int indent, const QStringList* preferredKeys = nullptr)
-{
-    QStringList keys;
-    if (preferredKeys) {
-        for (const QString& k : *preferredKeys) {
-            if (obj.contains(k))
-                keys.append(k);
-        }
-    }
-    for (auto it = obj.constBegin(); it != obj.constEnd(); ++it) {
-        if (!keys.contains(it.key()))
-            keys.append(it.key());
-    }
-    if (keys.isEmpty())
-        return "{}";
-
-    QString out = "{\n";
-    for (int i = 0; i < keys.size(); ++i) {
-        const QString& k = keys[i];
-        out += Indent(indent + 4) + "\"" + JsonEscaped(k) + "\": "
-               + SerializeJsonValueOrdered(obj.value(k), indent + 4, nullptr);
-        if (i + 1 < keys.size())
-            out += ",";
-        out += "\n";
-    }
-    out += Indent(indent) + "}";
-    return out;
-}
-
-QString SerializeJsonArrayOrdered(const QJsonArray& arr, int indent)
-{
-    if (arr.isEmpty())
-        return "[]";
-    QString out = "[\n";
-    for (int i = 0; i < arr.size(); ++i) {
-        out += Indent(indent + 4) + SerializeJsonValueOrdered(arr[i], indent + 4, nullptr);
-        if (i + 1 < arr.size())
-            out += ",";
-        out += "\n";
-    }
-    out += Indent(indent) + "]";
-    return out;
-}
-
-QString SerializeJsonValueOrdered(const QJsonValue& value, int indent, const QStringList* preferredKeys)
-{
-    if (value.isObject())
-        return SerializeJsonObjectOrdered(value.toObject(), indent, preferredKeys);
-    if (value.isArray())
-        return SerializeJsonArrayOrdered(value.toArray(), indent);
-    if (value.isString())
-        return QString("\"%1\"").arg(JsonEscaped(value.toString()));
-    if (value.isDouble()) {
-        const double d = value.toDouble();
-        return QString::number(d, 'g', 16);
-    }
-    if (value.isBool())
-        return value.toBool() ? "true" : "false";
-    if (value.isNull() || value.isUndefined())
-        return "null";
-    return "null";
-}
-
-QString SerializeJobDefRootPreservingTopOrder(const QJsonDocument& doc, const QString& rawText)
-{
-    if (doc.isArray()) {
-        const QJsonArray arr = doc.array();
-        QString out = "[\n";
-        for (int i = 0; i < arr.size(); ++i) {
-            const QStringList ordered = ExtractTopLevelJobDefKeys(rawText, QString::number(i));
-            out += Indent(4) + SerializeJsonValueOrdered(arr[i], 4, &ordered);
-            if (i + 1 < arr.size())
-                out += ",";
-            out += "\n";
-        }
-        out += "]";
-        return out;
-    }
-    if (doc.isObject()) {
-        const QStringList ordered = ExtractTopLevelJobDefKeys(rawText, "");
-        return SerializeJsonObjectOrdered(doc.object(), 0, &ordered);
-    }
-    return QString::fromUtf8(doc.toJson(QJsonDocument::Indented));
-}
-
-void CollectFormPathsInJsonOrder(const QJsonValue& value, const QString& basePath,
-                                 QStringList* scalarPaths, QStringList* arrayPaths)
-{
-    if (!scalarPaths || !arrayPaths)
-        return;
-    if (value.isObject()) {
-        const QJsonObject obj = value.toObject();
-        for (auto it = obj.constBegin(); it != obj.constEnd(); ++it) {
-            const QString childPath = basePath.isEmpty() ? it.key() : (basePath + "." + it.key());
-            CollectFormPathsInJsonOrder(it.value(), childPath, scalarPaths, arrayPaths);
-        }
-        return;
-    }
-    if (value.isArray()) {
-        const QJsonArray arr = value.toArray();
-        const JsonArrayShape shape = ClassifyArrayShape(arr);
-        if (shape == JsonArrayShape::Primitive) {
-            arrayPaths->append(basePath);
-            return;
-        }
-        if (shape == JsonArrayShape::FlatObject) {
-            for (int i = 0; i < arr.size(); ++i) {
-                const QJsonObject itemObj = arr[i].toObject();
-                for (auto it = itemObj.constBegin(); it != itemObj.constEnd(); ++it) {
-                    const QString p = QString("%1.%2.%3").arg(basePath).arg(i).arg(it.key());
-                    scalarPaths->append(p);
-                }
-            }
-        }
-        return;
-    }
-    scalarPaths->append(basePath);
-}
-
-QString EditTextOrEmpty(QLineEdit* edit)
-{
-    return edit ? edit->text() : QString();
-}
-
-QJsonValue ParseTextByType(const QString& txt, QJsonValue::Type type)
-{
-    if (type == QJsonValue::Bool)
-        return QJsonValue(txt.trimmed().toLower() == "true" || txt.trimmed() == "1");
-    if (type == QJsonValue::Double) {
-        bool ok = false;
-        const double d = txt.toDouble(&ok);
-        return ok ? QJsonValue(d) : QJsonValue(txt);
-    }
-    if (type == QJsonValue::Null)
-        return QJsonValue(QJsonValue::Null);
-    return QJsonValue(txt);
-}
-
-void ApplyScalarEditsInOrder(const QStringList& orderedScalarPaths,
-                             const QMap<QString, QLineEdit*>& scalarEdits,
-                             const QMap<QString, QJsonValue::Type>& scalarTypes,
-                             const std::function<void(const QString&, const QJsonValue&)>& applyValue)
-{
-    QSet<QString> applied;
-    for (const QString& path : orderedScalarPaths) {
-        if (!scalarEdits.contains(path))
-            continue;
-        QLineEdit* edit = scalarEdits.value(path, nullptr);
-        const QString text = edit ? edit->text() : QString();
-        const QJsonValue::Type type = scalarTypes.value(path, QJsonValue::String);
-        applyValue(path, ParseTextByType(text, type));
-        applied.insert(path);
-    }
-    for (auto it = scalarEdits.constBegin(); it != scalarEdits.constEnd(); ++it) {
-        const QString path = it.key();
-        if (applied.contains(path))
-            continue;
-        const QString text = it.value() ? it.value()->text() : QString();
-        const QJsonValue::Type type = scalarTypes.value(path, QJsonValue::String);
-        applyValue(path, ParseTextByType(text, type));
-    }
-}
-
-void ApplyPrimitiveArrayEditsInOrder(const QStringList& orderedArrayPaths,
-                                     const QMap<QString, QList<QLineEdit*>>& arrayEdits,
-                                     const QMap<QString, QJsonValue::Type>& arrayItemTypes,
-                                     const std::function<void(const QString&, const QJsonArray&)>& applyArray)
-{
-    QSet<QString> applied;
-    auto applyArrayPath = [&](const QString& path) {
-        if (!arrayEdits.contains(path))
-            return;
-        const QJsonValue::Type itemType = arrayItemTypes.value(path, QJsonValue::String);
-        QJsonArray arr;
-        for (QLineEdit* edit : arrayEdits.value(path)) {
-            if (!edit)
-                continue;
-            arr.append(ParseTextByType(edit->text(), itemType));
-        }
-        applyArray(path, arr);
-        applied.insert(path);
-    };
-    for (const QString& path : orderedArrayPaths)
-        applyArrayPath(path);
-    for (auto it = arrayEdits.constBegin(); it != arrayEdits.constEnd(); ++it) {
-        if (applied.contains(it.key()))
-            continue;
-        applyArrayPath(it.key());
-    }
-}
-
-bool ParseSubmitVariablesBlock(const QString& scriptText, QStringList* orderOut, QMap<QString, QString>* defaultsOut)
-{
-    if (!orderOut || !defaultsOut)
-        return false;
-    orderOut->clear();
-    defaultsOut->clear();
-
-    const QString marker = "########################################################################################";
-    const QString startTitle = "# Edit these variables for each run";
-    const QStringList lines = scriptText.split('\n');
-    int blockStart = -1;
-    for (int i = 0; i + 2 < lines.size(); ++i) {
-        if (lines[i].trimmed() == marker && lines[i + 1].trimmed() == startTitle && lines[i + 2].trimmed() == marker) {
-            blockStart = i + 3;
-            break;
-        }
-    }
-    if (blockStart < 0)
-        return false;
-    int blockEnd = lines.size();
-    for (int i = blockStart; i < lines.size(); ++i) {
-        if (lines[i].trimmed() == marker) {
-            blockEnd = i;
-            break;
-        }
-    }
-
-    QRegularExpression re("^\\s*([A-Za-z_][A-Za-z0-9_]*)\\s*=\\s*(.*)$");
-    QRegularExpression envDefaultRe("^\\$\\{[A-Za-z_][A-Za-z0-9_]*:-([^}]*)\\}$");
-    for (int i = blockStart; i < blockEnd; ++i) {
-        QString line = lines[i];
-        const int commentPos = line.indexOf('#');
-        if (commentPos >= 0)
-            line = line.left(commentPos);
-        line = line.trimmed();
-        if (line.isEmpty())
-            continue;
-        const QRegularExpressionMatch m = re.match(line);
-        if (!m.hasMatch())
-            continue;
-        const QString key = m.captured(1).trimmed();
-        QString value = m.captured(2).trimmed();
-        if (value.startsWith('"') && value.endsWith('"') && value.size() >= 2)
-            value = value.mid(1, value.size() - 2);
-        else if (value.startsWith('\'') && value.endsWith('\'') && value.size() >= 2)
-            value = value.mid(1, value.size() - 2);
-        const QRegularExpressionMatch envDefault = envDefaultRe.match(value);
-        if (envDefault.hasMatch()) {
-            value = envDefault.captured(1).trimmed();
-            if (value.startsWith('"') && value.endsWith('"') && value.size() >= 2)
-                value = value.mid(1, value.size() - 2);
-            else if (value.startsWith('\'') && value.endsWith('\'') && value.size() >= 2)
-                value = value.mid(1, value.size() - 2);
-        }
-        orderOut->append(key);
-        (*defaultsOut)[key] = value;
-    }
-    return !orderOut->isEmpty();
-}
-
-QString ShellDoubleQuote(const QString& value)
-{
-    QString out = value;
-    out.replace("\\", "\\\\");
-    out.replace("\"", "\\\"");
-    return "\"" + out + "\"";
-}
-
-bool BuildPatchedSubmitScript(const QString& scriptText, const QMap<QString, QString>& overrides, QString* patchedOut)
-{
-    if (!patchedOut)
-        return false;
-    const QString marker = "########################################################################################";
-    const QString startTitle = "# Edit these variables for each run";
-    QStringList lines = scriptText.split('\n');
-    int blockStart = -1;
-    for (int i = 0; i + 2 < lines.size(); ++i) {
-        if (lines[i].trimmed() == marker && lines[i + 1].trimmed() == startTitle && lines[i + 2].trimmed() == marker) {
-            blockStart = i + 3;
-            break;
-        }
-    }
-    if (blockStart < 0)
-        return false;
-    int blockEnd = lines.size();
-    for (int i = blockStart; i < lines.size(); ++i) {
-        if (lines[i].trimmed() == marker) {
-            blockEnd = i;
-            break;
-        }
-    }
-
-    QRegularExpression assignRe("^(\\s*)([A-Za-z_][A-Za-z0-9_]*)\\s*=\\s*([^#]*)(\\s*(?:#.*)?)$");
-    for (int i = blockStart; i < blockEnd; ++i) {
-        const QRegularExpressionMatch m = assignRe.match(lines[i]);
-        if (!m.hasMatch())
-            continue;
-        const QString key = m.captured(2).trimmed();
-        if (!overrides.contains(key))
-            continue;
-        const QString leading = m.captured(1);
-        QString suffix = m.captured(4);
-        if (suffix.startsWith('#'))
-            suffix.prepend(' ');
-        lines[i] = leading + key + "=" + ShellDoubleQuote(overrides.value(key)) + suffix;
-    }
-    const QString queueName = overrides.value("JOB_QUEUE").trimmed();
-    QString jobDefName = overrides.value("JOB_DEFINITION").trimmed();
-    if (jobDefName.isEmpty())
-        jobDefName = overrides.value("JOB_DEFINITION_NAME").trimmed();
-    for (int i = 0; i < lines.size(); ++i) {
-        QString line = lines[i];
-        if (line.trimmed().startsWith('#'))
-            continue;
-        if (!queueName.isEmpty()) {
-            QRegularExpression queueRe("^(\\s*--job-queue\\s+)([^\\s\\\\]+|\"[^\"]*\"|'[^']*')(\\s*\\\\?.*)$");
-            const QRegularExpressionMatch qm = queueRe.match(line);
-            if (qm.hasMatch()) {
-                lines[i] = qm.captured(1) + ShellDoubleQuote(queueName) + qm.captured(3);
-                continue;
-            }
-        }
-        if (!jobDefName.isEmpty()) {
-            QRegularExpression jdRe("^(\\s*--job-definition\\s+)([^\\s\\\\]+|\"[^\"]*\"|'[^']*')(\\s*\\\\?.*)$");
-            const QRegularExpressionMatch jm = jdRe.match(line);
-            if (jm.hasMatch()) {
-                lines[i] = jm.captured(1) + ShellDoubleQuote(jobDefName) + jm.captured(3);
-                continue;
-            }
-        }
-    }
-    *patchedOut = lines.join('\n');
-    return true;
-}
-
-QString ResolveSubmitValue(QString raw, const QMap<QString, QString>& values)
-{
-    if (raw.isEmpty())
-        return raw;
-    raw = raw.trimmed();
-    const QString today = QDate::currentDate().toString("yyyy-MM-dd");
-    raw.replace("$(date +%Y-%m-%d)", today);
-    raw.replace("$(date +%F)", today);
-
-    QRegularExpression bracedVarRe("\\$\\{([A-Za-z_][A-Za-z0-9_]*)\\}");
-    QRegularExpression plainVarRe("\\$([A-Za-z_][A-Za-z0-9_]*)");
-    for (int pass = 0; pass < 4; ++pass) {
-        bool changed = false;
-        QString next = raw;
-        QRegularExpressionMatchIterator it = bracedVarRe.globalMatch(next);
-        while (it.hasNext()) {
-            const QRegularExpressionMatch m = it.next();
-            const QString key = m.captured(1);
-            if (values.contains(key)) {
-                next.replace(m.captured(0), values.value(key));
-                changed = true;
-            }
-        }
-        it = plainVarRe.globalMatch(next);
-        while (it.hasNext()) {
-            const QRegularExpressionMatch m = it.next();
-            const QString key = m.captured(1);
-            if (values.contains(key)) {
-                next.replace(m.captured(0), values.value(key));
-                changed = true;
-            }
-        }
-        raw = next;
-        if (!changed)
-            break;
-    }
-    return raw;
-}
-
-} // namespace
+using namespace TsQtAWSUtils;
 
 // ---------------------------------------------------------------------------
 // Constructor
@@ -966,8 +93,8 @@ TsQtAWS::TsQtAWS(TsParameterManager* pm, QObject* parent)
       fJobDefEdit(nullptr), fEditJobDefSnapshotFolder(nullptr), fEditWizardJobQueueName(nullptr),
       fAwsDirEdit(nullptr), fProfileEdit(nullptr),
       fEditAwsRegion(nullptr), fBtnCancel(nullptr), fBtnReset(nullptr), fBtnNext(nullptr), fBtnDownloadS3(nullptr),
-      fBtnSubmitPostprocess(nullptr), fMonitorText(nullptr), fMonitorLastUpdated(nullptr), fPostprocessMonitorText(nullptr),
-      fPollTimer(nullptr), fPostTimer(nullptr), fEditProjectName(nullptr), fEditRunDate(nullptr),
+      fBtnSubmitPostprocess(nullptr), fMonitorText(nullptr), fMonitorLastUpdated(nullptr),
+      fPollTimer(nullptr), fEditProjectName(nullptr), fEditRunDate(nullptr),
       fEditNumJobs(nullptr), fEditJobName(nullptr), fEditInputBucket(nullptr), fEditOutputBucket(nullptr),
       fEditLocalSimDir(nullptr), fEditFileToRun(nullptr), fEditJobQueue(nullptr), fEditJobDefinitionName(nullptr),
       fEditPostprocessJobDefinitionName(nullptr), fEditPpLocalScript(nullptr), fEditPpExtraPip(nullptr), fEditDownloadDir(nullptr),
@@ -980,7 +107,7 @@ TsQtAWS::TsQtAWS(TsParameterManager* pm, QObject* parent)
       fJobBatchStepStack(nullptr), fRadioBatchExisting(nullptr), fRadioBatchCreate(nullptr),
       fRadioCreateUseExistingBucketsLogs(nullptr), fRadioCreateNewBucketsLogs(nullptr), fEditExistingInputBucket(nullptr),
       fEditExistingOutputBucket(nullptr), fEditExistingJobQueue(nullptr), fEditExistingJobDefSim(nullptr),
-      fEditExistingJobDefPost(nullptr), fChkExistingNoPostprocess(nullptr), fEditProvisionInputBucket(nullptr),
+      fEditExistingJobDefPost(nullptr), fEditProvisionInputBucket(nullptr),
       fEditProvisionOutputBucket(nullptr), fEditProvisionLogGroup(nullptr), fRadioLogGroupCreate(nullptr),
       fRadioLogGroupExisting(nullptr), fRadioSimJobDefExisting(nullptr), fRadioSimJobDefCreate(nullptr), fRadioProvisionPostYes(nullptr),
       fRadioProvisionPostNo(nullptr), fRadioPostJobDefExisting(nullptr), fRadioPostJobDefCreate(nullptr),
@@ -1051,1453 +178,7 @@ void TsQtAWS::ShowWizard(QWidget* parentWidget)
         "QLineEdit { border: 1px solid transparent; border-radius: 8px; padding: 2px 6px; }"
         "QLineEdit:focus { border: 2px solid rgb(10,132,255); border-radius: 8px; }");
 
-    QVBoxLayout* mainLay = new QVBoxLayout(fWizard);
-
-    QHBoxLayout* stepHeader = new QHBoxLayout();
-    fStepLabel = new QLabel("Welcome");
-    stepHeader->addWidget(fStepLabel, 1);
-    mainLay->addLayout(stepHeader);
-
-    fStepProgress = new QProgressBar();
-    fStepProgress->setRange(0, 9);
-    fStepProgress->setValue(0);
-    mainLay->addWidget(fStepProgress);
-
-    fStack = new QStackedWidget();
-
-    // Welcome page (before wizard step 1)
-    QWidget* pageWelcome = new QWidget();
-    pageWelcome->setFocusPolicy(Qt::ClickFocus);
-    QVBoxLayout* lWelcome = new QVBoxLayout(pageWelcome);
-    lWelcome->setContentsMargins(12, 18, 12, 18);
-    lWelcome->setSpacing(48);
-    auto loadTopasLogo = []() {
-        const QString appDir = QCoreApplication::applicationDirPath();
-        QStringList candidates;
-        candidates << QDir::homePath() + "/Applications/TOPAS/OpenTOPAS/graphics/TOPASLogo.png"
-                   << appDir + "/../../OpenTOPAS/graphics/TOPASLogo.png";
-        for (const QString& path : candidates) {
-            QPixmap pix(path);
-            if (!pix.isNull())
-                return pix;
-        }
-        return QPixmap();
-    };
-    QHBoxLayout* welcomeHeader = new QHBoxLayout();
-    welcomeHeader->setSpacing(14);
-    QLabel* logoLabel = new QLabel();
-    QPixmap logo = loadTopasLogo();
-    if (!logo.isNull())
-        logoLabel->setPixmap(logo.scaledToHeight(84, Qt::SmoothTransformation));
-    QLabel* welcomeTitle = new QLabel("<b>Welcome to the OpenTOPAS AWS Cloud Wizard!</b>");
-    welcomeTitle->setTextFormat(Qt::RichText);
-    QFont welcomeTitleFont = welcomeTitle->font();
-    welcomeTitleFont.setPointSize(welcomeTitleFont.pointSize() + 4);
-    welcomeTitle->setFont(welcomeTitleFont);
-    welcomeHeader->addWidget(logoLabel, 0, Qt::AlignVCenter);
-    welcomeHeader->addWidget(welcomeTitle, 1, Qt::AlignVCenter);
-    lWelcome->addLayout(welcomeHeader);
-    QLabel* welcomeText1 = new QLabel(
-        "This wizard will guide you through the process of setting up and running your simulations using AWS computing resources.");
-    welcomeText1->setWordWrap(true);
-    welcomeText1->setAlignment(Qt::AlignJustify);
-    lWelcome->addWidget(welcomeText1);
-    lWelcome->addSpacing(16);
-    QLabel* welcomeText2 = new QLabel(
-        "Before beginning these steps, please ensure that you have fully read the cloud section of the "
-        "<a href=\"https://topas-nbio.readthedocs.io/en/latest/\">documentation</a>. You will need to have a valid AWS account and have "
-        "completed up to and including <a href=\"https://topas-nbio.readthedocs.io/en/latest/\">step 4</a> of the setup instructions.");
-    welcomeText2->setTextFormat(Qt::RichText);
-    welcomeText2->setOpenExternalLinks(true);
-    welcomeText2->setWordWrap(true);
-    welcomeText2->setAlignment(Qt::AlignJustify);
-    lWelcome->addWidget(welcomeText2);
-    QLabel* supportText = new QLabel(
-        "If you encounter any problems using this wizard, please check the "
-        "<a href=\"https://github.com/OpenTOPAS/OpenTOPAS/discussions\">user forum</a> for relevant posts, or create your own post. "
-        "Please report any bugs by creating a GitHub "
-        "<a href=\"https://github.com/OpenTOPAS/OpenTOPAS/issues\">Issue</a>.");
-    supportText->setTextFormat(Qt::RichText);
-    supportText->setOpenExternalLinks(true);
-    supportText->setWordWrap(true);
-    supportText->setAlignment(Qt::AlignJustify);
-    lWelcome->addWidget(supportText);
-    QLabel* liabilityText = new QLabel(
-        "The TOPAS Collaboration is not responsible for any charges incurred while using AWS compute resources.");
-    liabilityText->setWordWrap(true);
-    lWelcome->addWidget(liabilityText);
-    lWelcome->addSpacing(16);
-    QPushButton* btnBegin = new QPushButton("Begin");
-    btnBegin->setSizePolicy(QSizePolicy::MinimumExpanding, QSizePolicy::Fixed);
-    btnBegin->setStyleSheet(
-        "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-        "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-        "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
-    lWelcome->addWidget(btnBegin);
-    lWelcome->addStretch(1);
-    fStack->addWidget(pageWelcome);
-
-    auto addUnderlinedTitle = [&](QVBoxLayout* layout, const QString& richText) {
-        layout->addSpacing(8);
-        QWidget* titleBlock = new QWidget();
-        QVBoxLayout* titleLayout = new QVBoxLayout(titleBlock);
-        titleLayout->setContentsMargins(0, 0, 0, 0);
-        titleLayout->setSpacing(4);
-        QLabel* title = new QLabel(richText);
-        title->setTextFormat(Qt::RichText);
-        QFont titleFont = title->font();
-        titleFont.setPointSize(titleFont.pointSize() + 3);
-        title->setFont(titleFont);
-        titleLayout->addWidget(title);
-        QFrame* underline = new QFrame();
-        underline->setFrameShape(QFrame::HLine);
-        underline->setFrameShadow(QFrame::Plain);
-        underline->setLineWidth(2);
-        underline->setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Fixed);
-        titleLayout->addWidget(underline);
-        layout->addWidget(titleBlock);
-        layout->addSpacing(8);
-    };
-
-    // Step 2 — default (field editors) vs advanced (JSON text)
-    QWidget* pageMode = new QWidget();
-    pageMode->setFocusPolicy(Qt::ClickFocus);
-    QVBoxLayout* lMode = new QVBoxLayout(pageMode);
-    lMode->setContentsMargins(12, 18, 12, 18);
-    lMode->setSpacing(16);
-    addUnderlinedTitle(lMode, "<b>Choose the type of setup instructions</b>");
-    QLabel* modeDesc = new QLabel(
-        "If you are a new user we recommend using the default setup steps. These steps will guide you through the "
-        "process of setting up your AWS environment based on the templates provided by TOPAS-nBio in the AWS "
-        "directory you downloaded.<br><br>"
-        "The advanced option allows you to directly edit the raw JSON or bash scripts for more fine-tuned control.");
-    modeDesc->setWordWrap(true);
-    lMode->addWidget(modeDesc);
-    fRadioWizardDefault = new QRadioButton("Default");
-    fRadioWizardAdvanced = new QRadioButton("Advanced");
-    fRadioWizardDefault->setChecked(true);
-    lMode->addWidget(fRadioWizardDefault);
-    lMode->addWidget(fRadioWizardAdvanced);
-    QObject::connect(fRadioWizardDefault, &QRadioButton::toggled, fWizard, [this](bool) { UpdateNavButtons(); });
-    QObject::connect(fRadioWizardAdvanced, &QRadioButton::toggled, fWizard, [this](bool) { UpdateNavButtons(); });
-    lMode->addStretch(1);
-
-    // Wizard step 1 of 7 — AWS scripts path, profile, and region (before workflow choice)
-    QWidget* page0 = new QWidget();
-    page0->setFocusPolicy(Qt::ClickFocus);
-    QVBoxLayout* l0 = new QVBoxLayout(page0);
-    l0->setContentsMargins(12, 18, 12, 18);
-    l0->setSpacing(16);
-    {
-        addUnderlinedTitle(l0, "<b>AWS scripts folder</b>");
-        QLabel* folderDetail = new QLabel(
-            "Provide the directory which contains the AWS Batch job definition and compute environment JSON files, "
-            "along with the simulation submission and post-processing Bash scripts. These files can be found "
-            "<a href=\"https://github.com/topas-nbio/TOPAS-nBio/tree/main/aws\">here</a> and should be copied to a local directory.<br><br>"
-            "This GUI is aimed at new users, who may not be familiar with the range of AWS services. As such, the current implementation "
-            "requires these exact files to be present. A more flexible solution may be implemented in the future.");
-        folderDetail->setTextFormat(Qt::RichText);
-        folderDetail->setOpenExternalLinks(true);
-        folderDetail->setWordWrap(true);
-        folderDetail->setAlignment(Qt::AlignJustify);
-        l0->addWidget(folderDetail);
-    }
-    l0->addSpacing(24);
-    fAwsDirEdit = new QLineEdit(fAwsDir);
-    fAwsDirEdit->setFixedHeight(22);
-    l0->addWidget(fAwsDirEdit);
-    QPushButton* browse = new QPushButton("Browse...");
-    browse->setSizePolicy(QSizePolicy::MinimumExpanding, QSizePolicy::Fixed);
-    browse->setStyleSheet(
-        "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-        "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-        "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
-    l0->addWidget(browse);
-    l0->addSpacing(36);
-    {
-        addUnderlinedTitle(l0, "<b>Environment variables</b>");
-    }
-    QLabel* envDescription = new QLabel(
-        "If you have correctly setup your AWS CLI profile, these environment variable fields should be pre-filled. "
-        "If not, please supply them.");
-    envDescription->setWordWrap(true);
-    l0->addWidget(envDescription);
-    l0->addSpacing(8);
-    QHBoxLayout* envRow = new QHBoxLayout();
-    envRow->setSpacing(16);
-    QVBoxLayout* profileCol = new QVBoxLayout();
-    {
-        QLabel* profileLbl = new QLabel("<b>AWS profile</b>");
-        profileLbl->setTextFormat(Qt::RichText);
-        profileCol->addWidget(profileLbl);
-    }
-    fProfileEdit = new QLineEdit(fProfile);
-    fProfileEdit->setFixedHeight(22);
-    profileCol->addWidget(fProfileEdit);
-    QVBoxLayout* regionCol = new QVBoxLayout();
-    {
-        QLabel* regionLbl = new QLabel("<b>AWS region</b>");
-        regionLbl->setTextFormat(Qt::RichText);
-        regionCol->addWidget(regionLbl);
-    }
-    fEditAwsRegion = new QLineEdit(fRegion);
-    fEditAwsRegion->setFixedHeight(22);
-    regionCol->addWidget(fEditAwsRegion);
-    envRow->addLayout(profileCol, 1);
-    envRow->addLayout(regionCol, 1);
-    l0->addLayout(envRow);
-    l0->addStretch(1);
-    fStack->addWidget(page0);
-    fStack->addWidget(pageMode);
-
-    QObject::connect(browse, &QPushButton::clicked, fWizard, [this]() {
-        QString d = QFileDialog::getExistingDirectory(fWizard, "AWS scripts folder", fAwsDirEdit->text());
-        if (!d.isEmpty()) {
-            fAwsDirEdit->setText(d);
-            ReloadAwsFilesFromUi();
-        }
-    });
-    QObject::connect(btnBegin, &QPushButton::clicked, fWizard, [this]() { SetStep(1); });
-
-    // Wizard step 3 — S3 buckets
-    QWidget* pageBuckets = new QWidget();
-    pageBuckets->setFocusPolicy(Qt::ClickFocus);
-    QVBoxLayout* lbk = new QVBoxLayout(pageBuckets);
-    lbk->setContentsMargins(12, 18, 12, 18);
-    lbk->setSpacing(12);
-    addUnderlinedTitle(lbk, "<b>S3 buckets</b>");
-    QLabel* bucketsDesc = new QLabel(
-        "See the <a href=\"https://topas-nbio.readthedocs.io/en/latest/\">relevant documentation</a> for more information about "
-        "AWS S3 buckets. "
-        "If you already have S3 buckets for TOPAS-nBio enter their names below. "
-        "Otherwise, provide names for the new buckets. These buckets will be created by this setup wizard and used in all future steps.");
-    bucketsDesc->setTextFormat(Qt::RichText);
-    bucketsDesc->setOpenExternalLinks(true);
-    bucketsDesc->setWordWrap(true);
-    lbk->addWidget(bucketsDesc);
-    fRadioCreateUseExistingBucketsLogs = new QRadioButton("Use existing buckets");
-    fRadioCreateNewBucketsLogs = new QRadioButton("Create new buckets");
-    fRadioCreateUseExistingBucketsLogs->setChecked(true);
-    lbk->addWidget(fRadioCreateUseExistingBucketsLogs);
-    lbk->addWidget(fRadioCreateNewBucketsLogs);
-    fEditProvisionInputBucket = new QLineEdit(QStringLiteral("topas-nbio-input"));
-    fEditProvisionInputBucket->setFixedHeight(22);
-    fEditProvisionInputBucket->setPlaceholderText(QStringLiteral("Input bucket name"));
-    lbk->addWidget(fEditProvisionInputBucket);
-    fEditProvisionOutputBucket = new QLineEdit(QStringLiteral("topas-nbio-output"));
-    fEditProvisionOutputBucket->setFixedHeight(22);
-    fEditProvisionOutputBucket->setPlaceholderText(QStringLiteral("Output bucket name"));
-    lbk->addWidget(fEditProvisionOutputBucket);
-    QObject::connect(fEditProvisionInputBucket, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    QObject::connect(fEditProvisionOutputBucket, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    QObject::connect(fRadioCreateNewBucketsLogs, &QRadioButton::toggled, fWizard, [this](bool) { UpdateNavButtons(); });
-    QObject::connect(fRadioCreateUseExistingBucketsLogs, &QRadioButton::toggled, fWizard, [this](bool) { UpdateNavButtons(); });
-    lbk->addStretch(1);
-    fStack->addWidget(pageBuckets);
-
-    // Wizard step 4 — compute environment (stack index 4): nested choice + editor
-    QWidget* page1 = new QWidget();
-    page1->setFocusPolicy(Qt::ClickFocus);
-    QVBoxLayout* l1Outer = new QVBoxLayout(page1);
-    l1Outer->setContentsMargins(0, 0, 0, 0);
-    fComputeStepStack = new QStackedWidget();
-    QWidget* computeChoicePage = new QWidget();
-    QVBoxLayout* lc = new QVBoxLayout(computeChoicePage);
-    lc->setContentsMargins(12, 18, 12, 18);
-    lc->setSpacing(14);
-    addUnderlinedTitle(lc, "<b>Setting up the compute environment</b>");
-    QLabel* computeChoiceDesc = new QLabel(
-        "See the <a href=\"https://topas-nbio.readthedocs.io/en/latest/\">relevant documentation</a> for more information about "
-        "AWS Batch compute environments. "
-        "If you have never launched simulations using AWS you should create a new compute "
-        "environment from the templates provided. Alternatively, if you have already launched simulations using AWS "
-        "you can use an existing compute environment by providing its name below.");
-    computeChoiceDesc->setTextFormat(Qt::RichText);
-    computeChoiceDesc->setOpenExternalLinks(true);
-    computeChoiceDesc->setWordWrap(true);
-    lc->addWidget(computeChoiceDesc);
-    fRadioComputeExisting = new QRadioButton("I already have a compute environment");
-    fRadioComputeCreate = new QRadioButton("Create a new compute environment from the template");
-    fRadioComputeExisting->setChecked(true);
-    lc->addWidget(fRadioComputeExisting);
-    fEditExistingComputeEnvName = new QLineEdit();
-    fEditExistingComputeEnvName->setFixedHeight(22);
-    fEditExistingComputeEnvName->setPlaceholderText("e.g. topas-nbio-compute-env");
-    lc->addWidget(fEditExistingComputeEnvName);
-    lc->addWidget(fRadioComputeCreate);
-    auto syncComputeNameEnabled = [this]() {
-        const bool ex = fRadioComputeExisting && fRadioComputeExisting->isChecked();
-        if (fEditExistingComputeEnvName)
-            fEditExistingComputeEnvName->setEnabled(ex);
-        if (fBtnNext)
-            UpdateNavButtons();
-    };
-    QObject::connect(fRadioComputeExisting, &QRadioButton::toggled, fWizard, [syncComputeNameEnabled](bool) { syncComputeNameEnabled(); });
-    QObject::connect(fRadioComputeCreate, &QRadioButton::toggled, fWizard, [syncComputeNameEnabled](bool) { syncComputeNameEnabled(); });
-    QObject::connect(fEditExistingComputeEnvName, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    lc->addStretch(1);
-    fComputeStepStack->addWidget(computeChoicePage);
-
-    QWidget* computeEditorPage = new QWidget();
-    QVBoxLayout* le = new QVBoxLayout(computeEditorPage);
-    le->setContentsMargins(0, 0, 0, 0);
-    fStep2Scroll = new QScrollArea();
-    fStep2Scroll->setWidgetResizable(true);
-    fStep2Scroll->setFrameShape(QFrame::NoFrame);
-    QWidget* step2Content = new QWidget();
-    QVBoxLayout* l1 = new QVBoxLayout(step2Content);
-    l1->setSpacing(18);
-    addUnderlinedTitle(l1, "<b>Setting up the compute environment</b>");
-    QLabel* computeDesc = new QLabel(
-        "Define the compute environment which determines the kind of resources AWS Batch uses. "
-        "Modify the template below, then continue to create the environment. "
-        "While some information is provided in the tooltip of each field, full information can be found in the "
-        "<a href=\"https://docs.aws.amazon.com/AWSCloudFormation/latest/TemplateReference/aws-resource-batch-computeenvironment.html\">AWS documentation</a>.");
-    computeDesc->setTextFormat(Qt::RichText);
-    computeDesc->setOpenExternalLinks(true);
-    computeDesc->setWordWrap(true);
-    l1->addWidget(computeDesc);
-    QLabel* computeDesc2 = new QLabel(
-        "If the pre-filled values are already correct, save the changes locally and proceed to create the compute environment. If they need to be modified "
-        "please refer to <a href=\"https://topas-nbio.readthedocs.io/en/latest/\"> section 6 </a> of the cloud documentation, which details "
-        "where exactly in the AWS console the values for the different fields can be found.");
-    computeDesc2->setTextFormat(Qt::RichText);
-    computeDesc2->setOpenExternalLinks(true);
-    computeDesc2->setWordWrap(true);
-    l1->addWidget(computeDesc2);
-    l1->addSpacing(14);
-    fComputeModeStack = new QStackedWidget();
-    fComputeFormPage = new QWidget();
-    QVBoxLayout* l1Form = new QVBoxLayout(fComputeFormPage);
-    l1Form->setContentsMargins(0, 0, 0, 0);
-    l1Form->setSpacing(15);
-    fComputeModeStack->addWidget(fComputeFormPage);
-    QWidget* computeTextPage = new QWidget();
-    QVBoxLayout* l1Text = new QVBoxLayout(computeTextPage);
-    l1Text->setContentsMargins(0, 0, 0, 0);
-    fComputeTextEditor = new QPlainTextEdit();
-    fComputeTextEditor->setFont(QFont("monospace"));
-    l1Text->addWidget(fComputeTextEditor);
-    QObject::connect(fComputeTextEditor, &QPlainTextEdit::textChanged, fWizard, [this]() {
-        if (fComputeTextMode)
-            MarkComputeNeedsApply();
-    });
-    fComputeEdit = fComputeTextEditor; // keep existing loading/apply wiring
-    fComputeModeStack->addWidget(computeTextPage);
-    l1->addWidget(fComputeModeStack);
-    addUnderlinedTitle(l1, "<b>Save changes locally</b>");
-    fComputeApplyDesc = new QLabel();
-    fComputeApplyDesc->setWordWrap(true);
-    l1->addWidget(fComputeApplyDesc);
-    l1->addWidget(new QLabel("Choose a folder name"));
-    fEditComputeSnapshotFolder = new QLineEdit();
-    fEditComputeSnapshotFolder->setFixedHeight(22);
-    l1->addWidget(fEditComputeSnapshotFolder);
-    QObject::connect(fEditComputeSnapshotFolder, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    l1->addStretch(1);
-    fStep2Scroll->setWidget(step2Content);
-    le->addWidget(fStep2Scroll);
-    fComputeStepStack->addWidget(computeEditorPage);
-    fComputeStepStack->setCurrentIndex(0);
-    l1Outer->addWidget(fComputeStepStack);
-    fComputeModeStack->setCurrentIndex(0);
-    fComputeTextMode = false;
-    fStack->addWidget(page1);
-
-    // Wizard step 5 — job queue, CloudWatch, and job definitions (stack index 5): nested flow
-    QWidget* page2 = new QWidget();
-    page2->setFocusPolicy(Qt::ClickFocus);
-    QVBoxLayout* l2OuterRoot = new QVBoxLayout(page2);
-    l2OuterRoot->setContentsMargins(0, 0, 0, 0);
-    fJobBatchStepStack = new QStackedWidget();
-
-    // --- 4a: job queue ---
-    QWidget* queueChoicePage = new QWidget();
-    QVBoxLayout* lbc = new QVBoxLayout(queueChoicePage);
-    lbc->setContentsMargins(12, 18, 12, 18);
-    lbc->setSpacing(12);
-    addUnderlinedTitle(lbc, "<b>Job queue</b>");
-    QLabel* batchChoiceDesc = new QLabel(
-        "See the <a href=\"https://topas-nbio.readthedocs.io/en/latest/\">relevant documentation</a> for more information about "
-        "AWS Batch job queues. "
-        "If you already have an AWS Batch job queue, choose the first option and enter its name below. "
-        "Otherwise choose create a new job queue (it can be created when you register job definitions).");
-    batchChoiceDesc->setTextFormat(Qt::RichText);
-    batchChoiceDesc->setOpenExternalLinks(true);
-    batchChoiceDesc->setWordWrap(true);
-    lbc->addWidget(batchChoiceDesc);
-    fRadioBatchExisting = new QRadioButton("I already have a job queue");
-    fRadioBatchCreate = new QRadioButton("Create a new job queue");
-    fRadioBatchExisting->setChecked(true);
-    lbc->addWidget(fRadioBatchExisting);
-    lbc->addWidget(fRadioBatchCreate);
-    QStackedWidget* queueNameStack = new QStackedWidget();
-    QWidget* existingQueueBlock = new QWidget();
-    QVBoxLayout* existingQueueLayout = new QVBoxLayout(existingQueueBlock);
-    existingQueueLayout->setContentsMargins(0, 0, 0, 0);
-    existingQueueLayout->setSpacing(0);
-    fEditExistingJobQueue = new QLineEdit();
-    fEditExistingJobQueue->setFixedHeight(22);
-    fEditExistingJobQueue->setPlaceholderText("e.g. topas-nbio-queue");
-    existingQueueLayout->addWidget(fEditExistingJobQueue);
-    queueNameStack->addWidget(existingQueueBlock);
-    QWidget* newQueueBlock = new QWidget();
-    QVBoxLayout* newQueueLayout = new QVBoxLayout(newQueueBlock);
-    newQueueLayout->setContentsMargins(0, 0, 0, 0);
-    newQueueLayout->setSpacing(0);
-    fEditWizardJobQueueName = new QLineEdit();
-    fEditWizardJobQueueName->setFixedHeight(22);
-    fEditWizardJobQueueName->setPlaceholderText("e.g. topas-nbio-queue");
-    newQueueLayout->addWidget(fEditWizardJobQueueName);
-    queueNameStack->addWidget(newQueueBlock);
-    lbc->addWidget(queueNameStack);
-    auto syncQueueChoiceFields = [this, queueNameStack]() {
-        const bool ex = fRadioBatchExisting && fRadioBatchExisting->isChecked();
-        if (queueNameStack)
-            queueNameStack->setCurrentIndex(ex ? 0 : 1);
-        if (fBtnNext)
-            UpdateNavButtons();
-    };
-    QObject::connect(fRadioBatchExisting, &QRadioButton::toggled, fWizard, [syncQueueChoiceFields](bool) { syncQueueChoiceFields(); });
-    QObject::connect(fRadioBatchCreate, &QRadioButton::toggled, fWizard, [syncQueueChoiceFields](bool) { syncQueueChoiceFields(); });
-    QObject::connect(fEditExistingJobQueue, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    QObject::connect(fEditWizardJobQueueName, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    syncQueueChoiceFields();
-    lbc->addStretch(1);
-    fJobBatchStepStack->addWidget(queueChoicePage);
-
-    // --- 5b: CloudWatch log group ---
-    QWidget* batchLogsPage = new QWidget();
-    QVBoxLayout* lblg = new QVBoxLayout(batchLogsPage);
-    lblg->setContentsMargins(12, 18, 12, 18);
-    lblg->setSpacing(12);
-    addUnderlinedTitle(lblg, "<b>CloudWatch log group</b>");
-    QLabel* logsDesc = new QLabel(
-        "See the <a href=\"https://topas-nbio.readthedocs.io/en/latest/\">relevant documentation</a> for more information about "
-        "Amazon CloudWatch Logs for AWS Batch. "
-        "If you already have a CloudWatch log group for Batch jobs, choose use existing and enter its name below. "
-        "Otherwise choose create new log group (the wizard can create it when provisioning).");
-    logsDesc->setTextFormat(Qt::RichText);
-    logsDesc->setOpenExternalLinks(true);
-    logsDesc->setWordWrap(true);
-    lblg->addWidget(logsDesc);
-    fRadioLogGroupExisting = new QRadioButton("Use an existing CloudWatch log group");
-    fRadioLogGroupCreate = new QRadioButton("Create a new CloudWatch log group");
-    fRadioLogGroupExisting->setChecked(true);
-    lblg->addWidget(fRadioLogGroupExisting);
-    lblg->addWidget(fRadioLogGroupCreate);
-    fEditProvisionLogGroup = new QLineEdit(QStringLiteral("/aws/batch/topas-nbio"));
-    fEditProvisionLogGroup->setFixedHeight(22);
-    fEditProvisionLogGroup->setPlaceholderText(QStringLiteral("Log group name"));
-    lblg->addWidget(fEditProvisionLogGroup);
-    QObject::connect(fEditProvisionLogGroup, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    QObject::connect(fRadioLogGroupCreate, &QRadioButton::toggled, fWizard, [this](bool) { UpdateNavButtons(); });
-    QObject::connect(fRadioLogGroupExisting, &QRadioButton::toggled, fWizard, [this](bool) { UpdateNavButtons(); });
-    lblg->addStretch(1);
-    fJobBatchStepStack->addWidget(batchLogsPage);
-
-    // --- 5c: simulation job definition source (existing vs create) ---
-    QWidget* simJobDefChoicePage = new QWidget();
-    QVBoxLayout* lsimChoice = new QVBoxLayout(simJobDefChoicePage);
-    lsimChoice->setContentsMargins(12, 18, 12, 18);
-    lsimChoice->setSpacing(14);
-    addUnderlinedTitle(lsimChoice, "<b>Setting up simulation job definition</b>");
-    QLabel* simChoiceDesc = new QLabel(
-        "See the <a href=\"https://topas-nbio.readthedocs.io/en/latest/\">relevant documentation</a> for more information about "
-        "AWS Batch job definitions for simulation. "
-        "If you have never launched simulations using AWS you should create a new simulation "
-        "job definition from the templates provided. Alternatively, if you have already launched simulations using AWS "
-        "you can use an existing simulation job definition by providing its name below.");
-    simChoiceDesc->setTextFormat(Qt::RichText);
-    simChoiceDesc->setOpenExternalLinks(true);
-    simChoiceDesc->setWordWrap(true);
-    lsimChoice->addWidget(simChoiceDesc);
-    fRadioSimJobDefExisting = new QRadioButton("I already have a simulation job definition");
-    fRadioSimJobDefCreate = new QRadioButton("Create a new simulation job definition from the template");
-    fRadioSimJobDefExisting->setChecked(true);
-    lsimChoice->addWidget(fRadioSimJobDefExisting);
-    fEditExistingJobDefSim = new QLineEdit();
-    fEditExistingJobDefSim->setFixedHeight(22);
-    fEditExistingJobDefSim->setPlaceholderText("e.g. topas-nbio-job");
-    lsimChoice->addWidget(fEditExistingJobDefSim);
-    lsimChoice->addWidget(fRadioSimJobDefCreate);
-    auto syncSimJobDefChoiceFields = [this]() {
-        const bool useExisting = fRadioSimJobDefExisting && fRadioSimJobDefExisting->isChecked();
-        fBatchUseExistingJobDefs = useExisting;
-        if (fEditExistingJobDefSim)
-            fEditExistingJobDefSim->setEnabled(useExisting);
-        if (fBtnNext)
-            UpdateNavButtons();
-    };
-    QObject::connect(fRadioSimJobDefExisting, &QRadioButton::toggled, fWizard, [syncSimJobDefChoiceFields](bool) { syncSimJobDefChoiceFields(); });
-    QObject::connect(fRadioSimJobDefCreate, &QRadioButton::toggled, fWizard, [syncSimJobDefChoiceFields](bool) { syncSimJobDefChoiceFields(); });
-    QObject::connect(fEditExistingJobDefSim, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    syncSimJobDefChoiceFields();
-    lsimChoice->addStretch(1);
-    fJobBatchStepStack->addWidget(simJobDefChoicePage);
-
-    // --- 5d: simulation job definition ---
-    QWidget* simJobDefPage = new QWidget();
-    QVBoxLayout* lsim = new QVBoxLayout(simJobDefPage);
-    lsim->setContentsMargins(12, 18, 12, 18);
-    lsim->setSpacing(12);
-    addUnderlinedTitle(lsim, "<b>Simulation job definition</b>");
-    QLabel* simJobDefDesc = new QLabel(
-        "Provide the simulation job definition name. The job definition determines how jobs on AWS should be run."
-            "Modify the template below, then continue to create the job definition. "
-            "While some information is provided in the tooltip of each field, full information can be found in the "
-            "<a href=\"https://docs.aws.amazon.com/AWSCloudFormation/latest/TemplateReference/aws-resource-batch-jobdefinition.html\">"
-            "AWS documentation</a>.");
-    simJobDefDesc->setTextFormat(Qt::RichText);
-    simJobDefDesc->setOpenExternalLinks(true);
-    simJobDefDesc->setWordWrap(true);
-    lsim->addWidget(simJobDefDesc);
-    QLabel* simJobDefDesc2 = new QLabel(
-        "If the pre-filled values are already correct, save the changes locally and proceed to create the job definition. If they need to be modified "
-        "please refer to <a href=\"https://topas-nbio.readthedocs.io/en/latest/\"> section 6 </a> of the cloud documentation, which details "
-        "where exactly in the AWS console the values for the different fields can be found.");
-    simJobDefDesc2->setTextFormat(Qt::RichText);
-    simJobDefDesc2->setOpenExternalLinks(true);
-    simJobDefDesc2->setWordWrap(true);
-    lsim->addWidget(simJobDefDesc2);
-    lsim->addSpacing(14);
-
-    QWidget* simHostWidget = new QWidget();
-    fSimJobDefHostLayout = new QVBoxLayout(simHostWidget);
-    fSimJobDefHostLayout->setContentsMargins(0, 0, 0, 0);
-    lsim->addWidget(simHostWidget, 1);
-
-    // --- 5e: post-processing (yes / no) ---
-    QWidget* batchPostPage = new QWidget();
-    QVBoxLayout* lbpp = new QVBoxLayout(batchPostPage);
-    lbpp->setContentsMargins(12, 18, 12, 18);
-    lbpp->setSpacing(12);
-    addUnderlinedTitle(lbpp, "<b>Post-processing</b>");
-    QLabel* postChoiceDesc = new QLabel(
-        "Do you want to post-process your results using this wizard as well? If so please ensure that you have a post-processing "
-        "script file in the same aws directory as all the other scripts. Please refer to <a href=\"https://topas-nbio.readthedocs.io/en/latest/\"> section 7 </a> " 
-        "of the cloud documentation for more information.");
-    postChoiceDesc->setWordWrap(true);
-    lbpp->addWidget(postChoiceDesc);
-    fRadioProvisionPostYes = new QRadioButton("Yes. Please select the script file you wish to use for post-processing.");
-    fRadioProvisionPostNo = new QRadioButton("No");
-    fRadioProvisionPostYes->setChecked(true);
-    lbpp->addWidget(fRadioProvisionPostYes);
-    lbpp->addWidget(fRadioProvisionPostNo);
-    QHBoxLayout* postScriptRow = new QHBoxLayout();
-    fEditPostprocessScriptPath = new QLineEdit();
-    fEditPostprocessScriptPath->setFixedHeight(22);
-    postScriptRow->addWidget(fEditPostprocessScriptPath, 1);
-    QPushButton* btnBrowsePostScript = new QPushButton("Browse...");
-    btnBrowsePostScript->setObjectName("postScriptBrowseBtn");
-    postScriptRow->addWidget(btnBrowsePostScript);
-    lbpp->addLayout(postScriptRow);
-    QObject::connect(btnBrowsePostScript, &QPushButton::clicked, fWizard, [this]() {
-        const QString base = (fAwsDirEdit && !fAwsDirEdit->text().trimmed().isEmpty()) ? fAwsDirEdit->text().trimmed() : fAwsDir;
-        const QString script = QFileDialog::getOpenFileName(fWizard, "Select post-processing script file", base);
-        if (!script.isEmpty() && fEditPostprocessScriptPath) {
-            fEditPostprocessScriptPath->setText(script);
-            fPostprocessScriptSelectedByBrowse = true;
-            UpdateNavButtons();
-        }
-    });
-    QObject::connect(fEditPostprocessScriptPath, &QLineEdit::textChanged, fWizard, [this](const QString&) {
-        fPostprocessScriptSelectedByBrowse = false;
-        UpdateNavButtons();
-    });
-    QObject::connect(fRadioProvisionPostYes, &QRadioButton::toggled, fWizard, [this, btnBrowsePostScript](bool y) {
-        if (y) {
-            fWantsPostprocessing = true;
-            if (fEditPostprocessScriptPath)
-                fEditPostprocessScriptPath->setEnabled(true);
-            if (btnBrowsePostScript)
-                btnBrowsePostScript->setEnabled(true);
-            UpdateJobDefTabVisibility();
-            UpdateNavButtons();
-        }
-    });
-    QObject::connect(fRadioProvisionPostNo, &QRadioButton::toggled, fWizard, [this, btnBrowsePostScript](bool y) {
-        if (y) {
-            fWantsPostprocessing = false;
-            if (fEditPostprocessScriptPath)
-                fEditPostprocessScriptPath->setEnabled(false);
-            if (btnBrowsePostScript)
-                btnBrowsePostScript->setEnabled(false);
-            UpdateJobDefTabVisibility();
-            UpdateNavButtons();
-        }
-    });
-    lbpp->addStretch(1);
-
-    // --- 5f: post-processing job definition source (existing vs create) ---
-    QWidget* postJobDefChoicePage = new QWidget();
-    QVBoxLayout* lpostChoice = new QVBoxLayout(postJobDefChoicePage);
-    lpostChoice->setContentsMargins(12, 18, 12, 18);
-    lpostChoice->setSpacing(14);
-    addUnderlinedTitle(lpostChoice, "<b>Setting up post-processing job definition</b>");
-    QLabel* postChoiceDesc2 = new QLabel(
-        "Choose whether to use an existing post-processing job definition or create a new one from the template.");
-    postChoiceDesc2->setWordWrap(true);
-    lpostChoice->addWidget(postChoiceDesc2);
-    fRadioPostJobDefExisting = new QRadioButton("I already have a post-processing job definition");
-    fRadioPostJobDefCreate = new QRadioButton("Create a new post-processing job definition from the template");
-    fRadioPostJobDefExisting->setChecked(true);
-    lpostChoice->addWidget(fRadioPostJobDefExisting);
-    fEditExistingJobDefPost = new QLineEdit();
-    fEditExistingJobDefPost->setFixedHeight(22);
-    fEditExistingJobDefPost->setPlaceholderText("e.g. topas-nbio-postprocess-job");
-    lpostChoice->addWidget(fEditExistingJobDefPost);
-    lpostChoice->addWidget(fRadioPostJobDefCreate);
-    auto syncPostJobDefChoiceFields = [this]() {
-        const bool useExisting = fRadioPostJobDefExisting && fRadioPostJobDefExisting->isChecked();
-        fBatchUseExistingPostJobDef = useExisting;
-        if (fEditExistingJobDefPost)
-            fEditExistingJobDefPost->setEnabled(useExisting);
-        if (fBtnNext)
-            UpdateNavButtons();
-    };
-    QObject::connect(fRadioPostJobDefExisting, &QRadioButton::toggled, fWizard, [syncPostJobDefChoiceFields](bool) { syncPostJobDefChoiceFields(); });
-    QObject::connect(fRadioPostJobDefCreate, &QRadioButton::toggled, fWizard, [syncPostJobDefChoiceFields](bool) { syncPostJobDefChoiceFields(); });
-    QObject::connect(fEditExistingJobDefPost, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    syncPostJobDefChoiceFields();
-    lpostChoice->addStretch(1);
-
-    // --- 5g: post-processing job definition ---
-    QWidget* postJobDefPage = new QWidget();
-    QVBoxLayout* lpost = new QVBoxLayout(postJobDefPage);
-    lpost->setContentsMargins(12, 18, 12, 18);
-    lpost->setSpacing(12);
-    addUnderlinedTitle(lpost, "<b>Post-processing job definition</b>");
-    QLabel* postJobDefDesc = new QLabel(
-        "Edit the post-processing job definition template. Bucket and log group values from earlier steps are reflected in the JSON.");
-    postJobDefDesc->setWordWrap(true);
-    lpost->addWidget(postJobDefDesc);
-    QWidget* postHostWidget = new QWidget();
-    fPostJobDefHostLayout = new QVBoxLayout(postHostWidget);
-    fPostJobDefHostLayout->setContentsMargins(0, 0, 0, 0);
-    lpost->addWidget(postHostWidget, 1);
-
-    // Shared editor (form / JSON) + snapshot folder — reparented between sim and post pages
-    fJobDefEditorHost = new QWidget();
-    QVBoxLayout* hlay = new QVBoxLayout(fJobDefEditorHost);
-    hlay->setContentsMargins(0, 0, 0, 0);
-    hlay->setSpacing(12);
-    fStep3Scroll = new QScrollArea();
-    fStep3Scroll->setWidgetResizable(true);
-    fStep3Scroll->setFrameShape(QFrame::NoFrame);
-    QWidget* step3Content = new QWidget();
-    QVBoxLayout* l2 = new QVBoxLayout(step3Content);
-    l2->setContentsMargins(0, 0, 0, 0);
-    l2->setSpacing(12);
-    fJobDefModeStack = new QStackedWidget();
-    fJobDefFormPage = new QWidget();
-    QVBoxLayout* l2Form = new QVBoxLayout(fJobDefFormPage);
-    l2Form->setContentsMargins(0, 0, 0, 0);
-    l2Form->setSpacing(15);
-    fJobDefModeStack->addWidget(fJobDefFormPage);
-    QWidget* jobDefTextPage = new QWidget();
-    QVBoxLayout* l2Text = new QVBoxLayout(jobDefTextPage);
-    l2Text->setContentsMargins(0, 0, 0, 0);
-    fJobDefTextEditor = new QPlainTextEdit();
-    fJobDefTextEditor->setFont(QFont("monospace"));
-    l2Text->addWidget(fJobDefTextEditor);
-    QObject::connect(fJobDefTextEditor, &QPlainTextEdit::textChanged, fWizard, [this]() {
-        if (fJobDefTextMode)
-            MarkJobDefNeedsApply();
-    });
-    fJobDefEdit = fJobDefTextEditor;
-    fJobDefModeStack->addWidget(jobDefTextPage);
-    l2->addWidget(fJobDefModeStack);
-    fStep3Scroll->setWidget(step3Content);
-    hlay->addWidget(fStep3Scroll, 1);
-    addUnderlinedTitle(hlay, "<b>Save changes locally</b>");
-    fJobDefApplyDesc = new QLabel();
-    fJobDefApplyDesc->setWordWrap(true);
-    hlay->addWidget(fJobDefApplyDesc);
-    hlay->addWidget(new QLabel("Choose a folder name"));
-    fEditJobDefSnapshotFolder = new QLineEdit();
-    fEditJobDefSnapshotFolder->setFixedHeight(22);
-    hlay->addWidget(fEditJobDefSnapshotFolder);
-    QObject::connect(fEditJobDefSnapshotFolder, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    fSimJobDefHostLayout->addWidget(fJobDefEditorHost, 1);
-    fJobDefModeStack->setCurrentIndex(0);
-    fJobDefTextMode = false;
-
-    fJobBatchStepStack->addWidget(simJobDefPage);
-    fJobBatchStepStack->addWidget(batchPostPage);
-    fJobBatchStepStack->addWidget(postJobDefChoicePage);
-    fJobBatchStepStack->addWidget(postJobDefPage);
-    fJobBatchStepStack->setCurrentIndex(0);
-    l2OuterRoot->addWidget(fJobBatchStepStack);
-    fStack->addWidget(page2);
-
-    // Wizard step 6 of 8 — cloud submission mode
-    QWidget* page3 = new QWidget();
-    page3->setFocusPolicy(Qt::ClickFocus);
-    QVBoxLayout* page3Outer = new QVBoxLayout(page3);
-    page3Outer->setContentsMargins(12, 18, 12, 18);
-    page3Outer->setSpacing(12);
-    addUnderlinedTitle(page3Outer, "<b>Cloud submission</b>");
-    QLabel* submitModeDesc = new QLabel(
-        "Do you want to submit the parameter file you opened and containing directory to AWS? "
-        "The directory should contain all includeFiles.");
-    submitModeDesc->setWordWrap(true);
-    page3Outer->addWidget(submitModeDesc);
-    fRadioCloudSubmitUseOpenParam = new QRadioButton("Yes, use the parameter file");
-    fRadioCloudSubmitUseTemplateVars = new QRadioButton("No, use the variables from the template script");
-    fRadioCloudSubmitUseOpenParam->setChecked(true);
-    page3Outer->addWidget(fRadioCloudSubmitUseOpenParam);
-    page3Outer->addWidget(fRadioCloudSubmitUseTemplateVars);
-    page3Outer->addStretch(1);
-    fStack->addWidget(page3);
-    QObject::connect(fRadioCloudSubmitUseOpenParam, &QRadioButton::toggled, fWizard, [this](bool checked) {
-        if (checked)
-            fSubmitUseOpenedParameterFile = true;
-        UpdateNavButtons();
-    });
-    QObject::connect(fRadioCloudSubmitUseTemplateVars, &QRadioButton::toggled, fWizard, [this](bool checked) {
-        if (checked)
-            fSubmitUseOpenedParameterFile = false;
-        UpdateNavButtons();
-    });
-
-    // Wizard step 7 of 8 — Batch submit parameters
-    QWidget* page4 = new QWidget();
-    page4->setFocusPolicy(Qt::ClickFocus);
-    QVBoxLayout* page4Outer = new QVBoxLayout(page4);
-    page4Outer->setContentsMargins(12, 18, 12, 18);
-    page4Outer->setSpacing(16);
-    addUnderlinedTitle(page4Outer, "<b>Cloud submission</b>");
-    QLabel* submitDesc = new QLabel("The fields below are editable variables from the submission script.");
-    submitDesc->setWordWrap(true);
-    page4Outer->addWidget(submitDesc);
-    fSubmitVarsGrid = new QGridLayout();
-    page4Outer->addLayout(fSubmitVarsGrid);
-    addUnderlinedTitle(page4Outer, "<b>Save changes locally</b>");
-    QLabel* submitApplyDesc = new QLabel(
-        "These settings will be saved locally in the AWS directory that you specified in step 1.");
-    submitApplyDesc->setWordWrap(true);
-    page4Outer->addWidget(submitApplyDesc);
-    page4Outer->addWidget(new QLabel("Choose a folder name"));
-    fEditSubmitSnapshotFolder = new QLineEdit();
-    fEditSubmitSnapshotFolder->setFixedHeight(22);
-    page4Outer->addWidget(fEditSubmitSnapshotFolder);
-    QObject::connect(fEditSubmitSnapshotFolder, &QLineEdit::textChanged, fWizard, [this](const QString&) { UpdateNavButtons(); });
-    page4Outer->addStretch(1);
-    fStack->addWidget(page4);
-
-    // Wizard step 8 of 8 — job monitoring
-    QWidget* page5 = new QWidget();
-    page5->setFocusPolicy(Qt::ClickFocus);
-    QVBoxLayout* l4 = new QVBoxLayout(page5);
-    l4->setContentsMargins(12, 18, 12, 18);
-    l4->setSpacing(14);
-    addUnderlinedTitle(l4, "<b>Job status dashboard</b>");
-    QLabel* monitorDesc1 = new QLabel(
-        "You can monitor the status of your submitted jobs on this page. It is refreshed periodically.");
-    monitorDesc1->setWordWrap(true);
-    l4->addWidget(monitorDesc1);
-    QLabel* monitorDesc2 = new QLabel(
-        "Once jobs are completed you will be able to click \"Next\" which will take you to the last step of this wizard "
-        "which allows you to download and optionally postprocess your results.");
-    monitorDesc2->setWordWrap(true);
-    l4->addWidget(monitorDesc2);
-    QLabel* monitorDesc3 = new QLabel(
-        "Exiting this monitoring page early will exit the TOPAS AWS Cloud wizard and you will need to download/postprocess "
-        "the simulation outputs yourself.");
-    monitorDesc3->setWordWrap(true);
-    l4->addWidget(monitorDesc3);
-    fMonitorLastUpdated = new QLabel("Dashboard last updated: --. Auto-refreshes every 60 seconds.");
-    l4->addWidget(fMonitorLastUpdated);
-
-    auto addStatusRow = [this, l4](const QString& key, const QString& labelText) {
-        QHBoxLayout* row = new QHBoxLayout();
-        row->setContentsMargins(0, 0, 0, 0);
-        row->setSpacing(10);
-        QFrame* dot = new QFrame();
-        dot->setFixedSize(12, 12);
-        dot->setStyleSheet("background-color: rgb(110,110,110); border-radius: 6px;");
-        QLabel* lbl = new QLabel(labelText);
-        row->addWidget(dot, 0, Qt::AlignVCenter);
-        row->addWidget(lbl, 0, Qt::AlignVCenter);
-        row->addStretch(1);
-        l4->addLayout(row);
-        fMonitorStateDots[key] = dot;
-        fMonitorStateLabels[key] = lbl;
-    };
-    addStatusRow("SUBMITTED", "0 jobs submitted");
-    addStatusRow("PENDING", "0 jobs pending");
-    addStatusRow("RUNNABLE", "0 jobs runnable");
-    addStatusRow("STARTING", "0 jobs starting");
-    addStatusRow("RUNNING", "0 jobs running");
-    addStatusRow("SUCCEEDED", "0 jobs succeeded");
-    addStatusRow("FAILED", "0 job/s failed");
-
-    fMonitorText = new QTextEdit();
-    fMonitorText->setReadOnly(true);
-    fMonitorText->setFont(QFont("monospace"));
-    fMonitorText->setVisible(false);
-    l4->addWidget(fMonitorText);
-    QPushButton* exitEarly = new QPushButton("Exit monitoring");
-    l4->addWidget(exitEarly);
-    fStack->addWidget(page5);
-    QObject::connect(exitEarly, &QPushButton::clicked, fWizard, [this]() {
-        QMessageBox m(fWizard);
-        m.setWindowTitle("Stop monitoring");
-        m.setText("If you exit now, you must monitor and post-process yourself outside this dialog.");
-        m.setIcon(QMessageBox::Warning);
-        m.setStandardButtons(QMessageBox::Ok | QMessageBox::Cancel);
-        if (m.exec() == QMessageBox::Ok)
-            SetStep(9); // download / post-process (skip wait on monitoring)
-    });
-
-    // Wizard step 9 of 9 — download and optional post-process
-    QWidget* page6 = new QWidget();
-    page6->setFocusPolicy(Qt::ClickFocus);
-    QVBoxLayout* l5 = new QVBoxLayout(page6);
-    l5->setContentsMargins(12, 18, 12, 18);
-    l5->setSpacing(8);
-    QLabel* downloadTitle = new QLabel("<b>Download outputs</b>");
-    downloadTitle->setTextFormat(Qt::RichText);
-    downloadTitle->setSizePolicy(QSizePolicy::Preferred, QSizePolicy::Fixed);
-    QFont downloadTitleFont = downloadTitle->font();
-    downloadTitleFont.setPointSize(downloadTitleFont.pointSize() + 3);
-    downloadTitle->setFont(downloadTitleFont);
-    l5->addWidget(downloadTitle);
-    QFrame* downloadUnderline = new QFrame();
-    downloadUnderline->setFrameShape(QFrame::HLine);
-    downloadUnderline->setFrameShadow(QFrame::Plain);
-    downloadUnderline->setLineWidth(2);
-    downloadUnderline->setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Fixed);
-    l5->addWidget(downloadUnderline);
-    QLabel* downloadDesc = new QLabel("Choose a local folder to download your data into.");
-    downloadDesc->setSizePolicy(QSizePolicy::Preferred, QSizePolicy::Fixed);
-    l5->addWidget(downloadDesc);
-    QHBoxLayout* h5 = new QHBoxLayout();
-    fEditDownloadDir = new QLineEdit();
-    fEditDownloadDir->setFixedHeight(22);
-    h5->addWidget(fEditDownloadDir);
-    QPushButton* browseDl = new QPushButton("Browse...");
-    h5->addWidget(browseDl);
-    l5->addLayout(h5);
-    fBtnDownloadS3 = new QPushButton("Download simulation job outputs from S3");
-    fBtnDownloadS3->setStyleSheet(
-        "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-        "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-        "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
-    l5->addWidget(fBtnDownloadS3);
-    l5->addSpacing(10);
-    fPostprocessSectionWidget = new QWidget();
-    QVBoxLayout* postSectionLayout = new QVBoxLayout(fPostprocessSectionWidget);
-    postSectionLayout->setContentsMargins(0, 0, 0, 0);
-    postSectionLayout->setSpacing(6);
-    QLabel* postTitle = new QLabel("<b>Post-process</b> (optional)");
-    postTitle->setTextFormat(Qt::RichText);
-    postTitle->setSizePolicy(QSizePolicy::Preferred, QSizePolicy::Fixed);
-    QFont postTitleFont = postTitle->font();
-    postTitleFont.setPointSize(postTitleFont.pointSize() + 3);
-    postTitle->setFont(postTitleFont);
-    postSectionLayout->addWidget(postTitle);
-    QFrame* postUnderline = new QFrame();
-    postUnderline->setFrameShape(QFrame::HLine);
-    postUnderline->setFrameShadow(QFrame::Plain);
-    postUnderline->setLineWidth(2);
-    postUnderline->setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Fixed);
-    postSectionLayout->addWidget(postUnderline);
-    fPostprocessVarsGrid = new QGridLayout();
-    postSectionLayout->addLayout(fPostprocessVarsGrid);
-    fBtnSubmitPostprocess = new QPushButton("Submit postprocess job");
-    fBtnSubmitPostprocess->setStyleSheet(
-        "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-        "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-        "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
-    postSectionLayout->addWidget(fBtnSubmitPostprocess);
-    l5->addWidget(fPostprocessSectionWidget);
-    l5->addStretch(1);
-    fStack->addWidget(page6);
-
-    QObject::connect(browseDl, &QPushButton::clicked, fWizard, [this]() {
-        QString d = QFileDialog::getExistingDirectory(fWizard, "Download folder", fEditDownloadDir->text());
-        if (!d.isEmpty())
-            fEditDownloadDir->setText(d);
-    });
-    QObject::connect(fBtnDownloadS3, &QPushButton::clicked, fWizard, [this]() {
-        const QString projectName = !fProjectName.isEmpty() ? fProjectName : EditTextOrEmpty(fEditProjectName);
-        const QString runDate = !fRunDate.isEmpty() ? fRunDate : EditTextOrEmpty(fEditRunDate);
-        const QString outBucket = !fOutputBucket.isEmpty() ? fOutputBucket : EditTextOrEmpty(fEditOutputBucket);
-        if (projectName.isEmpty() || runDate.isEmpty() || outBucket.isEmpty()) {
-            ShowAwsMessage(fWizardParent, "s3 sync", false, "Required submission variables are missing.");
-            return;
-        }
-        QString simDir = "projects/" + projectName + "/" + runDate;
-        QString s3uri = "s3://" + outBucket + "/" + simDir + "/";
-        QString localTarget = fEditDownloadDir->text();
-        if (fDownloadPostprocessResults)
-            s3uri += "postP_results/";
-        if (fDownloadPostprocessResults)
-            localTarget = QDir(localTarget).filePath("postP_results");
-        QStringList args;
-        args << "s3"
-             << "sync"
-             << s3uri
-             << localTarget
-             << "--only-show-errors";
-        QDialog* busy = new QDialog(fWizard);
-        busy->setWindowTitle("Download");
-        busy->setModal(true);
-        busy->setWindowFlag(Qt::WindowCloseButtonHint, false);
-        QVBoxLayout* dl = new QVBoxLayout(busy);
-        QLabel* status = new QLabel("Busy downloading ...");
-        status->setWordWrap(true);
-        dl->addWidget(status);
-        QHBoxLayout* buttons = new QHBoxLayout();
-        QPushButton* btnCancelDownload = new QPushButton("Cancel");
-        QPushButton* btnExitDownload = new QPushButton("Exit");
-        btnExitDownload->setEnabled(false);
-        btnCancelDownload->setStyleSheet(
-            "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-            "QPushButton:enabled { background-color: rgb(58,58,58); color: rgb(235,235,235); }"
-            "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
-        btnExitDownload->setStyleSheet(
-            "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-            "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-            "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
-        buttons->addWidget(btnCancelDownload);
-        buttons->addWidget(btnExitDownload);
-        buttons->addStretch(1);
-        dl->addLayout(buttons);
-
-        QProcess* proc = new QProcess(busy);
-        QStringList awsArgs;
-        AwsCliArgs(&awsArgs);
-        awsArgs.append(args);
-        QString* stdOut = new QString();
-        QString* stdErr = new QString();
-        bool* wasCancelled = new bool(false);
-        QObject::connect(busy, &QDialog::finished, busy, [stdOut, stdErr, wasCancelled](int) {
-            delete stdOut;
-            delete stdErr;
-            delete wasCancelled;
-        });
-        QObject::connect(proc, &QProcess::readyReadStandardOutput, busy, [proc, stdOut]() {
-            *stdOut += QString::fromUtf8(proc->readAllStandardOutput());
-        });
-        QObject::connect(proc, &QProcess::readyReadStandardError, busy, [proc, stdErr]() {
-            *stdErr += QString::fromUtf8(proc->readAllStandardError());
-        });
-        QObject::connect(btnCancelDownload, &QPushButton::clicked, busy, [proc, status, wasCancelled]() {
-            if (proc->state() == QProcess::NotRunning)
-                return;
-            *wasCancelled = true;
-            status->setText("Cancelling download ...");
-            proc->kill();
-        });
-        QObject::connect(btnExitDownload, &QPushButton::clicked, busy, [busy]() { busy->accept(); });
-        QObject::connect(proc, &QProcess::errorOccurred, busy, [status, btnCancelDownload, btnExitDownload](QProcess::ProcessError) {
-            status->setText("Download failed.");
-            btnCancelDownload->setEnabled(false);
-            btnExitDownload->setEnabled(true);
-        });
-        QObject::connect(proc, static_cast<void (QProcess::*)(int, QProcess::ExitStatus)>(&QProcess::finished), busy,
-                         [this, status, btnCancelDownload, btnExitDownload, stdOut, stdErr, wasCancelled]
-                         (int code, QProcess::ExitStatus st) {
-            btnCancelDownload->setEnabled(false);
-            btnExitDownload->setEnabled(true);
-            if (*wasCancelled) {
-                status->setText("Download cancelled.");
-                return;
-            }
-            if (st == QProcess::NormalExit && code == 0) {
-                status->setText("Download complete!");
-                return;
-            }
-            status->setText("Download failed.");
-            ShowAwsMessage(fWizardParent, "s3 sync", false, *stdErr + "\n" + *stdOut);
-        });
-        proc->start("aws", awsArgs);
-        if (!proc->waitForStarted(3000)) {
-            status->setText("Download failed.");
-            btnCancelDownload->setEnabled(false);
-            btnExitDownload->setEnabled(true);
-        }
-        busy->exec();
-        busy->deleteLater();
-    });
-    QObject::connect(fBtnSubmitPostprocess, &QPushButton::clicked, fWizard, [this]() {
-        fSnapshotRoot = SnapshotSubdir();
-        QDir().mkpath(fSnapshotRoot);
-        QString scriptPath = QDir(fSnapshotRoot).filePath("postprocess_gui.sh");
-        BuildPostprocessScriptToFile(scriptPath);
-        QString err;
-        int code = 0;
-        QString out;
-        if (!RunSubmitScript(scriptPath, &out, &err, &code) || code != 0) {
-            ShowAwsMessage(fWizardParent, "Postprocess", false, err + "\n" + out);
-            return;
-        }
-        QString line = out.trimmed().split('\n').last().trimmed();
-        fPostprocessJobId = line;
-        ShowAwsMessage(fWizardParent, "Postprocess", true, "Submitted job id: " + fPostprocessJobId);
-        fJobIds.clear();
-        if (!line.isEmpty())
-            fJobIds.append(line);
-        fMonitoringPostprocess = true;
-        fDownloadPostprocessResults = true;
-        SetStep(8);
-    });
-
-    mainLay->addWidget(fStack, 1);
-
-    QHBoxLayout* nav = new QHBoxLayout();
-    fBtnCancel = new QPushButton("Cancel");
-    fBtnReset = new QPushButton("Reset");
-    fBtnReset->setEnabled(false);
-    fBtnReset->setStyleSheet(
-        "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-        "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-        "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
-    fBtnNext = new QPushButton("Next");
-    nav->addWidget(fBtnCancel);
-    nav->addWidget(fBtnReset);
-    nav->addStretch(1);
-    nav->addWidget(fBtnNext);
-    mainLay->addLayout(nav);
-
-    QObject::connect(fBtnCancel, &QPushButton::clicked, fWizard, &QDialog::reject);
-    QObject::connect(fBtnReset, &QPushButton::clicked, fWizard, [this]() {
-        QMessageBox m(fWizard);
-        m.setWindowTitle("Reset values");
-        m.setIcon(QMessageBox::Warning);
-        m.setText("Are you sure you want to reset to default values? All changes will be lost.");
-        m.setStandardButtons(QMessageBox::Yes | QMessageBox::Cancel);
-        m.setDefaultButton(QMessageBox::Cancel);
-        if (m.exec() != QMessageBox::Yes)
-            return;
-
-        if (fCurrentStep == 3) {
-            if (fEditProvisionInputBucket) {
-                const QSignalBlocker b(fEditProvisionInputBucket);
-                fEditProvisionInputBucket->setText(QStringLiteral("topas-nbio-input"));
-            }
-            if (fEditProvisionOutputBucket) {
-                const QSignalBlocker b(fEditProvisionOutputBucket);
-                fEditProvisionOutputBucket->setText(QStringLiteral("topas-nbio-output"));
-            }
-            if (fRadioCreateNewBucketsLogs && fRadioCreateUseExistingBucketsLogs) {
-                const QSignalBlocker b1(fRadioCreateNewBucketsLogs);
-                const QSignalBlocker b2(fRadioCreateUseExistingBucketsLogs);
-                fRadioCreateUseExistingBucketsLogs->setChecked(true);
-                fRadioCreateNewBucketsLogs->setChecked(false);
-            }
-        } else if (fCurrentStep == 4) {
-            const int csi = ComputeStepStackIndex();
-            if (csi == 0) {
-                if (fRadioComputeExisting && fRadioComputeCreate) {
-                    const QSignalBlocker b1(fRadioComputeExisting);
-                    const QSignalBlocker b2(fRadioComputeCreate);
-                    fRadioComputeExisting->setChecked(true);
-                    fRadioComputeCreate->setChecked(false);
-                }
-                if (fEditExistingComputeEnvName) {
-                    const QSignalBlocker b(fEditExistingComputeEnvName);
-                    fEditExistingComputeEnvName->setText(fDefaultExistingComputeEnvName);
-                    fEditExistingComputeEnvName->setEnabled(true);
-                }
-            } else {
-                if (!fComputeDefaultJsonText.isEmpty() && fComputeTextEditor)
-                    fComputeTextEditor->setPlainText(fComputeDefaultJsonText);
-                fComputeJsonText = fComputeDefaultJsonText;
-                fComputeApplied = true;
-                fComputeNeedsApply = false;
-                fComputeCreated = false;
-                if (fEditComputeSnapshotFolder) {
-                    const QSignalBlocker b(fEditComputeSnapshotFolder);
-                    fEditComputeSnapshotFolder->clear();
-                }
-                RebuildComputeFormFromCurrentJson();
-            }
-        } else if (fCurrentStep == 5) {
-            const int jsi = JobBatchStepStackIndex();
-            if (jsi == 0) {
-                if (fEditExistingJobQueue) {
-                    const QSignalBlocker b(fEditExistingJobQueue);
-                    fEditExistingJobQueue->setText(fSubmitVarDefaults.value(QStringLiteral("JOB_QUEUE")));
-                }
-                if (fEditWizardJobQueueName) {
-                    const QSignalBlocker b(fEditWizardJobQueueName);
-                    fEditWizardJobQueueName->setText(fSubmitVarDefaults.value(QStringLiteral("JOB_QUEUE")));
-                }
-                if (fRadioBatchCreate && fRadioBatchExisting) {
-                    const QSignalBlocker b1(fRadioBatchCreate);
-                    const QSignalBlocker b2(fRadioBatchExisting);
-                    fRadioBatchExisting->setChecked(true);
-                    fRadioBatchCreate->setChecked(false);
-                }
-            } else if (jsi == 1) {
-                if (fEditProvisionLogGroup) {
-                    const QSignalBlocker b(fEditProvisionLogGroup);
-                    fEditProvisionLogGroup->setText(QStringLiteral("/aws/batch/topas-nbio"));
-                }
-                if (fRadioLogGroupCreate && fRadioLogGroupExisting) {
-                    const QSignalBlocker b1(fRadioLogGroupCreate);
-                    const QSignalBlocker b2(fRadioLogGroupExisting);
-                    fRadioLogGroupExisting->setChecked(true);
-                    fRadioLogGroupCreate->setChecked(false);
-                }
-            } else if (jsi == 2) {
-                fBatchUseExistingJobDefs = true;
-                if (fRadioSimJobDefCreate && fRadioSimJobDefExisting) {
-                    const QSignalBlocker b1(fRadioSimJobDefCreate);
-                    const QSignalBlocker b2(fRadioSimJobDefExisting);
-                    fRadioSimJobDefExisting->setChecked(true);
-                    fRadioSimJobDefCreate->setChecked(false);
-                }
-                if (fEditExistingJobDefSim) {
-                    const QSignalBlocker b(fEditExistingJobDefSim);
-                    QString jds = fSubmitVarDefaults.value(QStringLiteral("JOB_DEFINITION"));
-                    if (jds.isEmpty())
-                        jds = fSubmitVarDefaults.value(QStringLiteral("JOB_DEFINITION_NAME"));
-                    fEditExistingJobDefSim->setText(jds);
-                    fEditExistingJobDefSim->setEnabled(true);
-                }
-            } else if (jsi == 3) {
-                if (!fJobDefDefaultJsonText.isEmpty()) {
-                    fJobDefJsonText = fJobDefDefaultJsonText;
-                    fActiveJobDefTabIndex = 0;
-                    EnsureJobDefEditorHostForPage(true);
-                    LoadActiveJobDefIntoEditor();
-                }
-            } else if (jsi == 4) {
-                if (fRadioProvisionPostYes) {
-                    const QSignalBlocker b(fRadioProvisionPostYes);
-                    fRadioProvisionPostYes->setChecked(true);
-                }
-                if (fEditPostprocessScriptPath && fAwsDirEdit) {
-                    const QSignalBlocker b(fEditPostprocessScriptPath);
-                    fEditPostprocessScriptPath->setText(QDir::cleanPath(fAwsDirEdit->text().trimmed()));
-                    fEditPostprocessScriptPath->setEnabled(true);
-                }
-                if (fWizard) {
-                    if (QPushButton* browseBtn = fWizard->findChild<QPushButton*>("postScriptBrowseBtn"))
-                        browseBtn->setEnabled(true);
-                }
-                fPostprocessScriptSelectedByBrowse = false;
-            } else if (jsi == 5) {
-                if (fRadioPostJobDefExisting && fRadioPostJobDefCreate) {
-                    const QSignalBlocker b1(fRadioPostJobDefExisting);
-                    const QSignalBlocker b2(fRadioPostJobDefCreate);
-                    fRadioPostJobDefExisting->setChecked(true);
-                    fRadioPostJobDefCreate->setChecked(false);
-                }
-                if (fEditExistingJobDefPost) {
-                    const QSignalBlocker b(fEditExistingJobDefPost);
-                    fEditExistingJobDefPost->setText(QStringLiteral("topas-nbio-postprocess-job"));
-                    fEditExistingJobDefPost->setEnabled(true);
-                }
-            } else if (jsi == 6) {
-                if (!fJobDefPostDefaultJsonText.isEmpty()) {
-                    fJobDefPostJsonText = fJobDefPostDefaultJsonText;
-                    fActiveJobDefTabIndex = 1;
-                    EnsureJobDefEditorHostForPage(false);
-                    LoadActiveJobDefIntoEditor();
-                }
-            }
-            if (jsi >= 3) {
-                fJobDefApplied = true;
-                fJobDefNeedsApply = false;
-                fJobDefCreated = false;
-                RebuildJobDefFormFromCurrentJson();
-                if (fEditWizardJobQueueName) {
-                    const QSignalBlocker b(fEditWizardJobQueueName);
-                    fEditWizardJobQueueName->setText(fSubmitVarDefaults.value("JOB_QUEUE"));
-                }
-                if (fEditJobDefSnapshotFolder) {
-                    const QSignalBlocker b(fEditJobDefSnapshotFolder);
-                    fEditJobDefSnapshotFolder->clear();
-                }
-            }
-            if (fJobBatchStepStack)
-                fJobBatchStepStack->setCurrentIndex(jsi);
-        } else if (fCurrentStep == 7) {
-            for (const QString& key : fSubmitVarOrder) {
-                QLineEdit* edit = fSubmitVarEdits.value(key, nullptr);
-                if (edit)
-                    edit->setText(fSubmitVarDefaults.value(key));
-            }
-            fProjectName.clear();
-            fRunDate.clear();
-            fOutputBucket.clear();
-            fSubmitNeedsReset = false;
-            fSubmitApplied = true;
-            fSubmitScriptSnapshotPath = fTopasSubmitPath;
-            if (fEditSubmitSnapshotFolder) {
-                const QSignalBlocker b(fEditSubmitSnapshotFolder);
-                fEditSubmitSnapshotFolder->clear();
-            }
-        }
-        UpdateNavButtons();
-    });
-    QObject::connect(fBtnNext, &QPushButton::clicked, fWizard, [this]() {
-        if (fCurrentStep == 9) { // final step — close
-            fWizard->accept();
-            return;
-        }
-        if (fCurrentStep == 2) { // default vs advanced
-            SaveAwsSessionOnly();
-            fAwsDir = fAwsDirEdit ? fAwsDirEdit->text() : fAwsDir;
-            if (!HasRequiredFiles(fAwsDir)) {
-                ShowAwsMessage(fWizardParent, "AWS scripts folder", false,
-                               "The folder must contain the TOPAS-nBio template files. Go back to step 1 to select the folder.");
-                return;
-            }
-            SetWizardEditorModesFromStep2();
-            SetStep(3);
-            return;
-        }
-        if (fCurrentStep == 3) { // buckets
-            const QString inB = fEditProvisionInputBucket ? fEditProvisionInputBucket->text().trimmed() : QString();
-            const QString outB = fEditProvisionOutputBucket ? fEditProvisionOutputBucket->text().trimmed() : QString();
-            if (inB.isEmpty() || outB.isEmpty()) {
-                ShowAwsMessage(fWizardParent, "S3 buckets", false, "Enter both input and output bucket names.");
-                return;
-            }
-            SetStep(4);
-            if (fComputeStepStack)
-                fComputeStepStack->setCurrentIndex(0);
-            return;
-        }
-        if (fCurrentStep == 4) { // compute: nested choice or editor
-            const int csi = ComputeStepStackIndex();
-            if (csi == 0) {
-                const bool wantExisting = fRadioComputeExisting && fRadioComputeExisting->isChecked();
-                if (wantExisting) {
-                    const QString n = fEditExistingComputeEnvName ? fEditExistingComputeEnvName->text().trimmed() : QString();
-                    if (n.isEmpty()) {
-                        ShowAwsMessage(fWizardParent, "Compute environment", false, "Enter the name of your existing compute environment.");
-                        return;
-                    }
-                    fComputeUseExisting = true;
-                    fExistingComputeEnvName = n;
-                    fComputeCreated = true;
-                    fComputeApplied = true;
-                    SetStep(5);
-                    if (fJobBatchStepStack)
-                        fJobBatchStepStack->setCurrentIndex(0);
-                    return;
-                }
-                fComputeUseExisting = false;
-                fExistingComputeEnvName.clear();
-                if (fComputeStepStack)
-                    fComputeStepStack->setCurrentIndex(1);
-                UpdateNavButtons();
-                return;
-            }
-            if (!fComputeCreated) {
-                if (!CreateComputeEnvironmentFromSnapshot())
-                    return;
-                fComputeCreated = true;
-            }
-            SetStep(5);
-            if (fJobBatchStepStack)
-                fJobBatchStepStack->setCurrentIndex(0);
-            return;
-        }
-        if (fCurrentStep == 5) { // batch / job defs nested
-            const int jsi = JobBatchStepStackIndex();
-            if (jsi == 0) {
-                const bool useExisting = fRadioBatchExisting && fRadioBatchExisting->isChecked();
-                const QString jq = useExisting
-                    ? (fEditExistingJobQueue ? fEditExistingJobQueue->text().trimmed() : QString())
-                    : (fEditWizardJobQueueName ? fEditWizardJobQueueName->text().trimmed() : QString());
-                if (jq.isEmpty()) {
-                    ShowAwsMessage(fWizardParent, "Job queue", false, "Please enter a job queue name.");
-                    return;
-                }
-                fBatchUseExisting = useExisting;
-                if (fJobBatchStepStack)
-                    fJobBatchStepStack->setCurrentIndex(1);
-                UpdateNavButtons();
-                return;
-            }
-            if (jsi == 1) {
-                QString provErr;
-                if (!RunProvisionS3AndLogs(&provErr)) {
-                    ShowAwsMessage(fWizardParent, "Provision S3 / logs", false, provErr);
-                    return;
-                }
-                fS3LogsProvisioned = true;
-                const QString inB = fEditProvisionInputBucket ? fEditProvisionInputBucket->text().trimmed() : QString();
-                const QString outB = fEditProvisionOutputBucket ? fEditProvisionOutputBucket->text().trimmed() : QString();
-                const QString lg = fEditProvisionLogGroup ? fEditProvisionLogGroup->text().trimmed() : QString();
-                const QString reg = fEditAwsRegion ? fEditAwsRegion->text().trimmed() : QString();
-                fJobDefJsonText = PatchJobDefJsonBucketsAndLog(fJobDefJsonText, inB, outB, lg, reg);
-                fJobDefPostJsonText = PatchJobDefJsonBucketsAndLog(fJobDefPostJsonText, inB, outB, lg, reg);
-                fActiveJobDefTabIndex = 0;
-                EnsureJobDefEditorHostForPage(true);
-                UpdateJobDefTabVisibility();
-                LoadActiveJobDefIntoEditor();
-                if (!fJobDefTextMode)
-                    RebuildJobDefFormFromCurrentJson();
-                fJobDefApplied = true;
-                fJobDefCreated = false;
-                fJobDefNeedsApply = false;
-                if (fJobBatchStepStack)
-                    fJobBatchStepStack->setCurrentIndex(2);
-                UpdateNavButtons();
-                return;
-            }
-            if (jsi == 2) {
-                if (fBatchUseExistingJobDefs) {
-                    const QString existingSimJobDef = fEditExistingJobDefSim ? fEditExistingJobDefSim->text().trimmed() : QString();
-                    if (existingSimJobDef.isEmpty()) {
-                        ShowAwsMessage(fWizardParent, "Simulation job definition", false,
-                                       "Enter the name of your existing simulation job definition.");
-                        return;
-                    }
-                    if (fSubmitVarEdits.contains(QStringLiteral("JOB_DEFINITION")))
-                        fSubmitVarEdits[QStringLiteral("JOB_DEFINITION")]->setText(existingSimJobDef);
-                    if (fSubmitVarEdits.contains(QStringLiteral("JOB_DEFINITION_NAME")))
-                        fSubmitVarEdits[QStringLiteral("JOB_DEFINITION_NAME")]->setText(existingSimJobDef);
-                    if (fJobBatchStepStack)
-                        fJobBatchStepStack->setCurrentIndex(4);
-                    UpdateNavButtons();
-                    return;
-                }
-                if (!fSimJobDefPageInitialized && !fJobDefDefaultJsonText.isEmpty()) {
-                    fJobDefJsonText = fJobDefDefaultJsonText;
-                    fActiveJobDefTabIndex = 0;
-                    EnsureJobDefEditorHostForPage(true);
-                    LoadActiveJobDefIntoEditor();
-                    if (!fJobDefTextMode)
-                        RebuildJobDefFormFromCurrentJson();
-                    fSimJobDefPageInitialized = true;
-                }
-                if (fJobBatchStepStack)
-                    fJobBatchStepStack->setCurrentIndex(3);
-                UpdateNavButtons();
-                return;
-            }
-            if (jsi == 3) {
-                CommitJobDefEditorToActiveString();
-                if (fJobBatchStepStack)
-                    fJobBatchStepStack->setCurrentIndex(4);
-                UpdateNavButtons();
-                return;
-            }
-            if (jsi == 4) {
-                fWantsPostprocessing = fRadioProvisionPostYes && fRadioProvisionPostYes->isChecked();
-                UpdateJobDefTabVisibility();
-                if (!fWantsPostprocessing) {
-                    if (!fJobDefCreated) {
-                        if (!CreateJobDefinitionFromSnapshot())
-                            return;
-                        fJobDefCreated = true;
-                    }
-                    SyncSubmitStepFromBatchState();
-                    SetStep(6);
-                    return;
-                }
-                const QString localPostScript = fEditPostprocessScriptPath ? fEditPostprocessScriptPath->text().trimmed() : QString();
-                if (localPostScript.isEmpty()) {
-                    ShowAwsMessage(fWizardParent, "Post-processing", false,
-                                   "Please select the post-processing script file before continuing.");
-                    return;
-                }
-                if (!QFileInfo(localPostScript).exists() || !QFileInfo(localPostScript).isFile()) {
-                    ShowAwsMessage(fWizardParent, "Post-processing", false,
-                                   "Please select a valid post-processing script file before continuing.");
-                    return;
-                }
-                if (fEditPpLocalScript)
-                    fEditPpLocalScript->setText(localPostScript);
-                if (fJobBatchStepStack)
-                    fJobBatchStepStack->setCurrentIndex(5);
-                UpdateNavButtons();
-                return;
-            }
-            if (jsi == 5) {
-                const bool useExistingPost = fRadioPostJobDefExisting && fRadioPostJobDefExisting->isChecked();
-                fBatchUseExistingPostJobDef = useExistingPost;
-                if (useExistingPost) {
-                    const QString existingPostJobDef = fEditExistingJobDefPost ? fEditExistingJobDefPost->text().trimmed() : QString();
-                    if (existingPostJobDef.isEmpty()) {
-                        ShowAwsMessage(fWizardParent, "Post-processing job definition", false,
-                                       "Enter the name of your existing post-processing job definition.");
-                        return;
-                    }
-                    if (fEditPostprocessJobDefinitionName)
-                        fEditPostprocessJobDefinitionName->setText(existingPostJobDef);
-                    if (!fJobDefCreated) {
-                        if (!CreateJobDefinitionFromSnapshot())
-                            return;
-                        fJobDefCreated = true;
-                    }
-                    SyncSubmitStepFromBatchState();
-                    SetStep(6);
-                    return;
-                }
-                fActiveJobDefTabIndex = 1;
-                if (!fPostJobDefPageInitialized && !fJobDefPostDefaultJsonText.isEmpty()) {
-                    fJobDefPostJsonText = fJobDefPostDefaultJsonText;
-                    fPostJobDefPageInitialized = true;
-                }
-                EnsureJobDefEditorHostForPage(false);
-                LoadActiveJobDefIntoEditor();
-                if (!fJobDefTextMode)
-                    RebuildJobDefFormFromCurrentJson();
-                if (fJobBatchStepStack)
-                    fJobBatchStepStack->setCurrentIndex(6);
-                UpdateNavButtons();
-                return;
-            }
-            if (jsi == 6) {
-                if (!fJobDefCreated) {
-                    if (!CreateJobDefinitionFromSnapshot())
-                        return;
-                    fJobDefCreated = true;
-                }
-                SyncSubmitStepFromBatchState();
-                SetStep(6);
-                return;
-            }
-        }
-        if (fCurrentStep == 6) { // cloud submission choice -> submit page
-            SetStep(7);
-            return;
-        }
-        if (fCurrentStep == 7) { // submit then go to monitoring
-            if (fTopasSubmitPath.isEmpty() || !QFileInfo::exists(fTopasSubmitPath)) {
-                ShowAwsMessage(fWizardParent, "Submit to cloud", false, "Could not find topas_submit.sh in the AWS directory.");
-                return;
-            }
-            QMap<QString, QString> submitEnv = BuildSubmitOverridesFromUi();
-            if (!ApplySubmitScriptSnapshot(false))
-                return;
-            const QString submitScriptPath =
-                fSubmitScriptSnapshotPath.isEmpty() ? fTopasSubmitPath : fSubmitScriptSnapshotPath;
-            QString out;
-            QString err;
-            int code = 0;
-            if (!RunSubmitScript(submitScriptPath, &out, &err, &code, submitEnv) || code != 0) {
-                ShowAwsMessage(fWizardParent, "Submit to cloud", false, err + "\n" + out);
-                return;
-            }
-            fProjectName = ResolveSubmitValue(submitEnv.value("PROJECT_NAME"), submitEnv);
-            fRunDate = ResolveSubmitValue(submitEnv.value("RUN_DATE"), submitEnv);
-            fOutputBucket = ResolveSubmitValue(submitEnv.value("OUTPUT_BUCKET"), submitEnv);
-            ParseJobIdsFromSubmitOutput(out);
-            fMonitoringPostprocess = false;
-            fDownloadPostprocessResults = false;
-            ShowAwsMessage(fWizardParent, "Submit to cloud", true, "Submitted. Job ids:\n" + out);
-            SetStep(8); // monitoring
-            return;
-        }
-        SetStep(fCurrentStep + 1);
-    });
-
-    QObject::connect(fAwsDirEdit, &QLineEdit::editingFinished, fWizard, [this]() { ReloadAwsFilesFromUi(); });
-    QObject::connect(fProfileEdit, &QLineEdit::editingFinished, fWizard, [this]() { SaveAwsSessionOnly(); });
-    QObject::connect(fEditAwsRegion, &QLineEdit::editingFinished, fWizard, [this]() { SaveAwsSessionOnly(); });
+    CreateWizardPagesAndNavigation();
 
     fAwsDirEdit->setText(fAwsDir);
     fProfileEdit->setText(fProfile);
@@ -2508,7 +189,6 @@ void TsQtAWS::ShowWizard(QWidget* parentWidget)
     fWizard->exec();
 
     StopMonitoring();
-    StopPostprocessMonitor();
 
     delete fWizard;
     fWizard = nullptr;
@@ -2543,7 +223,6 @@ void TsQtAWS::ShowWizard(QWidget* parentWidget)
     fEditExistingJobQueue = nullptr;
     fEditExistingJobDefSim = nullptr;
     fEditExistingJobDefPost = nullptr;
-    fChkExistingNoPostprocess = nullptr;
     fRadioPostJobDefExisting = nullptr;
     fRadioPostJobDefCreate = nullptr;
     fRadioCloudSubmitUseOpenParam = nullptr;
@@ -2567,7 +246,6 @@ void TsQtAWS::ShowWizard(QWidget* parentWidget)
     fMonitorLastUpdated = nullptr;
     fMonitorStateDots.clear();
     fMonitorStateLabels.clear();
-    fPostprocessMonitorText = nullptr;
     fEditProjectName = nullptr;
     fEditRunDate = nullptr;
     fEditNumJobs = nullptr;
@@ -2786,10 +464,7 @@ void TsQtAWS::UpdateNavButtons()
         fBtnNext->setEnabled(true);
     }
     if (fCurrentStep > 0) {
-        fBtnNext->setStyleSheet(
-            "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-            "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-            "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
+        ApplyPrimaryButtonStyle(fBtnNext);
     } else {
         fBtnNext->setStyleSheet("");
     }
@@ -3018,10 +693,7 @@ void TsQtAWS::PromptAddArrayValue(const QString& path)
     QDialogButtonBox* box = new QDialogButtonBox();
     QPushButton* addBtn = box->addButton("Add", QDialogButtonBox::AcceptRole);
     addBtn->setEnabled(false);
-    addBtn->setStyleSheet(
-        "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-        "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-        "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
+    ApplyPrimaryButtonStyle(addBtn);
     box->addButton("Cancel", QDialogButtonBox::RejectRole);
     layout->addWidget(box);
     QObject::connect(input, &QLineEdit::textChanged, &dialog, [addBtn](const QString& t) { addBtn->setEnabled(!t.trimmed().isEmpty()); });
@@ -3057,10 +729,7 @@ void TsQtAWS::PromptRemoveArrayValue(const QString& path)
     QDialogButtonBox* box = new QDialogButtonBox();
     QPushButton* removeBtn = box->addButton("Remove", QDialogButtonBox::AcceptRole);
     removeBtn->setEnabled(combo->count() > 0);
-    removeBtn->setStyleSheet(
-        "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-        "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-        "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
+    ApplyPrimaryButtonStyle(removeBtn);
     box->addButton("Cancel", QDialogButtonBox::RejectRole);
     layout->addWidget(box);
     QObject::connect(box, &QDialogButtonBox::accepted, &dialog, &QDialog::accept);
@@ -3332,10 +1001,7 @@ void TsQtAWS::PromptAddJobDefArrayValue(const QString& path)
     QDialogButtonBox* box = new QDialogButtonBox();
     QPushButton* addBtn = box->addButton("Add", QDialogButtonBox::AcceptRole);
     addBtn->setEnabled(false);
-    addBtn->setStyleSheet(
-        "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-        "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-        "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
+    ApplyPrimaryButtonStyle(addBtn);
     box->addButton("Cancel", QDialogButtonBox::RejectRole);
     layout->addWidget(box);
     QObject::connect(input, &QLineEdit::textChanged, &dialog, [addBtn](const QString& t) { addBtn->setEnabled(!t.trimmed().isEmpty()); });
@@ -3369,10 +1035,7 @@ void TsQtAWS::PromptRemoveJobDefArrayValue(const QString& path)
     QDialogButtonBox* box = new QDialogButtonBox();
     QPushButton* removeBtn = box->addButton("Remove", QDialogButtonBox::AcceptRole);
     removeBtn->setEnabled(combo->count() > 0);
-    removeBtn->setStyleSheet(
-        "QPushButton { border-radius: 8px; padding: 6px 12px; border: none; }"
-        "QPushButton:enabled { background-color: rgb(10,132,255); color: white; }"
-        "QPushButton:disabled { background-color: rgb(58,58,58); color: rgb(136,136,136); }");
+    ApplyPrimaryButtonStyle(removeBtn);
     box->addButton("Cancel", QDialogButtonBox::RejectRole);
     layout->addWidget(box);
     QObject::connect(box, &QDialogButtonBox::accepted, &dialog, &QDialog::accept);
@@ -4756,63 +2419,5 @@ void TsQtAWS::StopMonitoring()
         fPollTimer->stop();
 }
 
-// ---------------------------------------------------------------------------
-// Postprocess job polling
-// ---------------------------------------------------------------------------
-
-void TsQtAWS::RefreshPostprocessStatus()
-{
-    if (fPostprocessJobId.isEmpty()) {
-        fPostprocessMonitorText->setPlainText("No postprocess job id.");
-        return;
-    }
-    QStringList args;
-    args << "batch"
-         << "describe-jobs"
-         << "--jobs" << fPostprocessJobId;
-    QString out;
-    QString err;
-    int code = 0;
-    if (!RunAws(args, &out, &err, &code) || code != 0) {
-        fPostprocessMonitorText->setPlainText(QString("describe-jobs failed:\n") + err);
-        return;
-    }
-    fPostprocessMonitorText->setPlainText(out);
-}
-
-void TsQtAWS::OnPollPostprocessJob()
-{
-    RefreshPostprocessStatus();
-    QString t = fPostprocessMonitorText->toPlainText();
-    QJsonParseError pe;
-    QJsonDocument jd = QJsonDocument::fromJson(t.toUtf8(), &pe);
-    if (pe.error != QJsonParseError::NoError || !jd.isObject())
-        return;
-    QJsonArray jobs = jd.object().value("jobs").toArray();
-    bool anyRunning = false;
-    for (const QJsonValue& jv : jobs) {
-        QString st = jv.toObject().value("status").toString();
-        if (st == "RUNNING" || st == "PENDING" || st == "RUNNABLE" || st == "STARTING")
-            anyRunning = true;
-    }
-    if (!anyRunning && !jobs.isEmpty() && fPostTimer)
-        fPostTimer->stop();
-}
-
-void TsQtAWS::StartPostprocessMonitor()
-{
-    if (!fPostTimer) {
-        fPostTimer = new QTimer(this);
-        connect(fPostTimer, &QTimer::timeout, this, &TsQtAWS::OnPollPostprocessJob);
-    }
-    RefreshPostprocessStatus();
-    fPostTimer->start(8000);
-}
-
-void TsQtAWS::StopPostprocessMonitor()
-{
-    if (fPostTimer)
-        fPostTimer->stop();
-}
 
 #endif
